% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_helpers.R
\name{ascii_clean}
\alias{ascii_clean}
\title{Force a string to be ASCII}
\usage{
ascii_clean(s)
}
\arguments{
\item{s}{(\code{character}) string(s) to clean}
}
\value{
(\code{character}) cleaned string(s)
}
\description{
Force a string to be ASCII
}
