% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_chimera.R
\name{bimera_denovo_table}
\alias{bimera_denovo_table}
\alias{bimera_denovo_table.matrix}
\alias{bimera_denovo_table.data.frame}
\title{Generate a table of denovo bimeras}
\usage{
bimera_denovo_table(
  seqtab,
  seqs = NULL,
  minFoldParentOverAbundance = 1.5,
  minParentAbundance = 2,
  allowOneOff = FALSE,
  minOneOffParentDistance = 4,
  maxShift = 16,
  multithread = FALSE,
  ...
)

\method{bimera_denovo_table}{matrix}(
  seqtab,
  seqs = colnames(seqtab),
  minFoldParentOverAbundance = 1.5,
  minParentAbundance = 2,
  allowOneOff = FALSE,
  minOneOffParentDistance = 4,
  maxShift = 16,
  multithread = FALSE,
  ...
)

\method{bimera_denovo_table}{data.frame}(
  seqtab,
  seqs = NULL,
  minFoldParentOverAbundance = 1.5,
  minParentAbundance = 2,
  allowOneOff = FALSE,
  minOneOffParentDistance = 4,
  maxShift = 16,
  multithread = FALSE,
  ...
)
}
\arguments{
\item{seqtab}{(\code{matrix} or \code{data.frame}) input sequence table, as returned by
\code{dada2::makeSequenceTable()}, \code{make_long_sequence_table()}, or
\code{make_mapped_sequence_table()}}

\item{seqs}{(file name of a FASTA file, \code{character} vector,
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, or \code{NULL}) sequences to use
in the table. If \code{NULL}, the sequences must be included in \code{seqtab}, either
as column names (for the \code{matrix} method) or as a column named \code{seq} (for the
\code{data.frame} method).}

\item{minFoldParentOverAbundance}{(\code{numeric(1)}) see \code{dada2::removeBimeraDenovo()}}

\item{minParentAbundance}{(\code{numeric(1)}) see \code{dada2::removeBimeraDenovo()}}

\item{allowOneOff}{(\code{logical(1)}) see \code{dada2::removeBimeraDenovo()}}

\item{minOneOffParentDistance}{(\code{numeric(1)}) see \code{dada2::removeBimeraDenovo()}}

\item{maxShift}{(\code{numeric(1)}) see \code{dada2::removeBimeraDenovo()}}

\item{multithread}{(\code{logical(1)}) if \code{TRUE}, use multiple threads for the
computation. If \code{FALSE}, use a single thread. If an integer, use that many
threads. If \code{NULL}, use the default number of threads.}

\item{...}{additional arguments to be passed to the method}
}
\value{
a \code{data.frame} with one of the columns \code{seq}, \code{seq_id}, or \code{seq_idx},
as well as columns \code{nsam}, and \code{nflag}, where \code{seq}
is the full sequence, \code{seq_id} is a string identifier, and \code{seq_idx} is an
integer identifier, \code{nsam} is the number of samples in which the sequence
appears, and \code{nflag} is the number of samples in which the sequence is
flagged as a bimera.
}
\description{
This function performs the per-sample \emph{de novo} chimera detection step of
\code{dada2::isBimeraDenovoTable()}, but does not do the consensus step. This
allows the function to be performed on smaller partitions of the sequence
table to reduce total memory usage. The results can then be combined using
\code{combine_bimera_denovo_tables()}. The implementation calls the internal API
of DADA2 and so it may be fragile to changes in the DADA2 package.
}
