% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{build_taxonomy}
\alias{build_taxonomy}
\alias{build_taxonomy_new}
\title{Combine tip classifications to build a full PROTAX taxonomy}
\usage{
build_taxonomy(...)

build_taxonomy_new(...)
}
\arguments{
\item{...}{(\code{character} vectors) the classifications to combine; these should
be comma-delimited classifications from most inclusive rank to least
inclusive.}
}
\value{
a \code{data.frame} with columns \code{taxon_id}, \code{parent_id}, \code{rank},
\code{classification}, and \code{prior}, as required to write a Protax "\code{taxonomy}" file
}
\description{
Combine tip classifications to build a full PROTAX taxonomy
}
