% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{calc_taxon_thresholds}
\alias{calc_taxon_thresholds}
\title{Calculate clustering thresholds for each taxon, falling back to its ancestor
taxa as necessary}
\usage{
calc_taxon_thresholds(
  rank,
  conf_level,
  taxon_table,
  fmeasure_optima,
  default = ingroup_taxon()
)
}
\arguments{
\item{rank}{(\code{character} string) the rank within which clustering will be
performed}

\item{conf_level}{(\code{character} string) as in \code{fmeasure_optima}}

\item{taxon_table}{(\code{data.frame}) taxonomy table; column "seq_id" gives the
sequence ID, and columns {\code{root_rank()}} to {\code{tip_rank()}} (e.g., "kingdom"
to "species") give the taxonomy at each rank.}

\item{fmeasure_optima}{(\code{data.frame}) optimum clustering thresholds within
various taxa; column "rank" gives the rank which is approximated by
clustering; "superrank" gives the rank of the taxon within which the
clustering threshold was optimized; "supertaxon" gives that taxon name;
"conf_level" gives a string description of the confidence level threshold for
taxonomic assignments; "threshold" gives the optimum clustering threshold;
"f_measure" gives the F measure at the optimum threshold.}

\item{default}{(\verb{character string}) default taxon to define threshold to use
when taxonomy is unknown. default: \code{ingroup_taxon()}}
}
\value{
(named \code{numeric}, where names are taxa and values are clustering
thresholds)
unknown sequences, grouped by taxonomy at the parent rank. Sequences where
the parent rank is also unknown are in the item named \code{"_NA_"}.
}
\description{
Calculate clustering thresholds for each taxon, falling back to its ancestor
taxa as necessary
}
