% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{collapseNoMismatch_vsearch}
\alias{collapseNoMismatch_vsearch}
\title{Collapse ASVs which are identical in their overlaps using VSEARCH}
\usage{
collapseNoMismatch_vsearch(seqtab, ..., ncpu = local_cpus())
}
\arguments{
\item{seqtab}{(integer \code{matrix} ) DADA2-style sequence table as returned by
\code{dada2::makeSequenceTable()}}

\item{...}{additional arguments to pass to \code{vsearch_cluster_smallmem()}}

\item{ncpu}{(\code{integer} count) number of threads to use}
}
\value{
\code{integer} matrix with identical sequences collapsed
}
\description{
Collapse ASVs which are identical in their overlaps using VSEARCH
}
