% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_chimera.R
\name{combine_bimera_denovo_tables}
\alias{combine_bimera_denovo_tables}
\title{Combine bimera tables from \code{bimera_denovo_table()}}
\usage{
combine_bimera_denovo_tables(
  bimdf,
  minSampleFraction = 0.9,
  ignoreNNegatives = 1L,
  verbose = FALSE
)
}
\arguments{
\item{bimdf}{(\code{data.frame}) output of one or more calls to
\code{bimera_denovo_table()}}

\item{minSampleFraction}{(\code{numeric(1)}) see \code{dada2::isBimeraDenovoTable()}}

\item{ignoreNNegatives}{(\code{integer(1)}) see \code{dada2::isBimeraDenovoTable()}}

\item{verbose}{(\code{logical(1)}) see \code{dada2::isBimeraDenovoTable()}}
}
\value{
if \code{bimdf} was generated from a
\link[dada2:makeSequenceTable]{dada2-style sequence table} or a
\link[=make_long_sequence_table]{long sequence table} with an explicit \code{seq} column,
then a character vector of sequences that are bimeras. If \code{bimdf} was a
\link[=make_mapped_sequence_table]{mapped sequence table} with a \code{seq_id} or
\code{seq_idx} column, then values corresponding to the \code{seq_id} or \code{seq_idx} of
the bimeras.
}
\description{
Combine bimera tables from \code{bimera_denovo_table()}
}
