% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\name{cutadapt_filter_trim}
\alias{cutadapt_filter_trim}
\title{Run cutadapt on single-end reads}
\usage{
cutadapt_filter_trim(
  file,
  primer,
  trim,
  options = cutadapt_options(),
  ncpu = local_cpus(),
  cutadapt = find_cutadapt(),
  logfile = NULL,
  ...
)
}
\arguments{
\item{file}{(\code{character}) path to the input file}

\item{primer}{(\code{character}) primer sequence}

\item{trim}{(\code{character}) path to the trimmed output file}

\item{options}{(\href{`cutadapt_options()`}{\code{cutadapt_options}}) options for
cutadapt}

\item{ncpu}{(\code{integer}) number of CPUs to use}

\item{cutadapt}{(\code{character}) path to the cutadapt executable}

\item{logfile}{(\code{character}) path to a log file which will receive the
cutadapt stdout.}

\item{...}{additional arguments; currently ignored}
}
\value{
a \code{character} string giving the trimmed output file
}
\description{
Run cutadapt on single-end reads
}
