% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\docType{data}
\name{cutadapt_option_names}
\alias{cutadapt_option_names}
\alias{cutadapt_options}
\title{Options for cutadapt single-read filtering and trimming}
\format{
An object of class \code{character} of length 11.
}
\usage{
cutadapt_option_names

cutadapt_options(
  max_err = 0.2,
  min_overlap = 10L,
  action = "retain",
  discard_untrimmed = TRUE,
  max_n = 0L,
  max_ee = NULL,
  min_length = NULL,
  max_length = NULL,
  truncQ = NULL,
  cut = NULL,
  compression_level = 6
)
}
\arguments{
\item{max_err}{(\code{numeric}) maximum error rate for adapter matching}

\item{min_overlap}{(\code{integer}) minimum overlap between read and adapter}

\item{action}{(\code{character}) action to take on the read; options are "trim",
"retain", "mask", "lowercase", "none"}

\item{discard_untrimmed}{(\code{logical}) discard reads that do not match the
adapter}

\item{max_n}{(\code{integer}) maximum number of Ns allowed in a read after trimming}

\item{max_ee}{(\code{numeric}) maximum expected errors in a read after trimming}

\item{min_length}{(\code{integer}) minimum length of a read after trimming}

\item{max_length}{(\code{integer}) maximum length of a read after trimming}

\item{truncQ}{(\code{integer}) quality score threshold for truncating the 3' end
of the read; or a vector of two integers for the 5' and 3' ends}

\item{cut}{(\code{integer}) number of bases to unconditionally cut from the 5' (if
positive) or 3' (if negative) end of the read prior to adapter trimming; or
a vector of two integers with different signs for both ends}

\item{compression_level}{(\code{integer}) compression level for the output file;
options are 1-9}
}
\value{
an object of class \code{cutadapt_options}
}
\description{
Options for cutadapt single-read filtering and trimming
}
\keyword{datasets}
