% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\name{cutadapt_paired_filter_trim}
\alias{cutadapt_paired_filter_trim}
\title{Run cutadapt on paired-end reads}
\usage{
cutadapt_paired_filter_trim(
  file_R1,
  file_R2,
  primer_R1,
  primer_R2,
  trim_R1,
  trim_R2,
  options = cutadapt_paired_options(),
  ncpu = local_cpus(),
  cutadapt = find_cutadapt(),
  logfile = NULL,
  ...
)
}
\arguments{
\item{file_R1}{(\code{character}) path to the R1 input file}

\item{file_R2}{(\code{character}) path to the R2 input file}

\item{primer_R1}{(\code{character}) primer sequence for R1}

\item{primer_R2}{(\code{character}) primer sequence for R2}

\item{trim_R1}{(\code{character}) path to the trimmed R1 output file}

\item{trim_R2}{(\code{character}) path to the trimmed R2 output file}

\item{options}{(\href{`cutadapt_paired_options()`}{\code{cutadapt_paired_options}})
options for cutadapt}

\item{ncpu}{(\code{integer}) number of CPUs to use}

\item{cutadapt}{(\code{character}) path to the cutadapt executable}

\item{logfile}{(\code{character}) path to a log file which will receive the
cutadapt stdout.}

\item{...}{additional arguments; currently ignored}
}
\value{
a character vector of the trimmed output files
}
\description{
Run cutadapt on paired-end reads
}
