% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\name{cutadapt_paired_option_names}
\alias{cutadapt_paired_option_names}
\alias{cutadapt_paired_options}
\title{Options for cutadapt filtering and trimming}
\usage{
cutadapt_paired_option_names

cutadapt_paired_options(
  max_err = 0.2,
  min_overlap = 10L,
  action = "retain",
  discard_untrimmed = TRUE,
  max_n = 0L,
  max_ee = NULL,
  min_length = NULL,
  max_length = NULL,
  truncQ_R1 = 2,
  truncQ_R2 = 2,
  cut_R1 = NULL,
  cut_R2 = NULL
)
}
\arguments{
\item{max_err}{(\code{numeric}) maximum error rate for adapter matching}

\item{min_overlap}{(\code{integer}) minimum overlap between read and adapter}

\item{action}{(\code{character}) action to take on the read; options are "trim",
"retain", "mask", "lowercase", "none"}

\item{discard_untrimmed}{(\code{logical}) discard reads that do not match the
adapter}

\item{max_n}{(\code{integer}) maximum number of Ns allowed in a read}

\item{max_ee}{(\code{numeric}) maximum expected errors in a read after trimming}

\item{min_length}{(\code{integer}) minimum length of a read after trimming}

\item{max_length}{(\code{integer}) maximum length of a read after trimming}

\item{truncQ_R1}{(\code{integer}) quality score threshold for truncating the 3'
end of read 1; or a vector of two integers for the 5' and 3' ends}

\item{truncQ_R2}{(\code{integer}) quality score threshold for truncating the 3'
end of read 2; or a vector of two integers for the 5' and 3' ends}

\item{cut_R1}{(\code{integer}) number of bases to unconditionally cut from the 5'
(if positive) or 3' (if negative) end of read 1 prior to adapter trimming;
or a vector of two integers with different signs for both ends}

\item{cut_R2}{(\code{integer}) number of bases to unconditinoally cut from the 5'
(if positive) or 3' (if negative) end of read 2 prior to adapter trimming;
or a vector of two integers with different signs for both ends}
}
\value{
an object of class \code{cutadapt_paired_options}
}
\description{
Options for cutadapt filtering and trimming

Options for cutadapt paired-end filtering and trimming
}
