% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{dada}
\alias{dada}
\title{Wrapper for dada2::dada}
\usage{
dada(
  derep,
  err,
  errorEstimationFunction = choose_dada_error_function,
  selfConsist = FALSE,
  pool = FALSE,
  priors = character(0),
  multithread = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{derep}{Dereplicated reads}

\item{err}{Error object}

\item{errorEstimationFunction}{Function to estimate error rates}

\item{selfConsist}{Use self-consistency}

\item{pool}{Pool reads}

\item{priors}{Sequences with prior information}

\item{multithread}{Use multiple threads}

\item{verbose}{Print progress messages}

\item{...}{Additional arguments (passed to \code{\link[dada2:dada]{dada2::dada()}})}
}
\value{
A \code{list} of \code{\link[dada2:dada-class]{dada}} objects, or \code{NULL} if no
\code{errorEstimationFunction} is given and \code{selfConsist} is \code{FALSE}.  If only 0
or 1 \code{\link[dada2:derep-class]{derep}} objects are given, the output will still
be a \code{list}.
}
\description{
Wrapper for dada2::dada
}
