% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{dada2_filter_options}
\alias{dada2_filter_options}
\title{Class to store dada2 filter options}
\usage{
dada2_filter_options(maxEE_R1, maxEE_R2)
}
\arguments{
\item{maxEE_R1}{Maximum expected errors for forward reads}

\item{maxEE_R2}{Maximum expected errors for reverse reads}
}
\value{
A list with class "dada2_filter_options"
}
\description{
Class to store dada2 filter options
}
