% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_map.R
\name{dada_merge_map}
\alias{dada_merge_map}
\title{Map the fate of individual reads through dada2 dereplication, denoising, and merge.}
\usage{
dada_merge_map(dadaF, derepF, dadaR, derepR, merged)
}
\arguments{
\item{dadaF}{(\code{\link[dada2:dada-class]{dada2::dada}} object or list of such objects) denoised
forward reads}

\item{derepF}{(\code{\link[dada2:derep-class]{dada2::derep}} object or list of such objects)
dereplicated forward reads}

\item{dadaR}{(\code{\link[dada2:dada-class]{dada2::dada}} object or list of such objects) denoised
reverse reads}

\item{derepR}{(\code{\link[dada2:derep-class]{dada2::derep}} object or list of such objects)
dereplicated reverse reads}

\item{merged}{(\code{data.frame} returned by \code{dada2::mergePairs()} or list of
such objects) results of merginf the denoised reads in dadaF and dadaR}
}
\value{
a \code{data.frame} with three columns:
\itemize{
\item \code{fwd_idx} (integer) index of forward ASV in \code{dadaF}
\item \code{rev_idx} (integer) index of reverse ASV in \code{dadaR}
\item \code{merge_idx} (integer) row index of merged ASV \code{merged}
Each row of this \code{data.frame} represents a single read in the fastq files
originally passed to \code{dada2::derepFastq()}, and the rows are in the same
order as the reads.
If the inputs were lists, then the output is a list of \code{data.frame}s as
described above.
}
}
\description{
Map the fate of individual reads through dada2 dereplication, denoising, and merge.
}
