% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table.R
\name{deduplicate_seq_idx}
\alias{deduplicate_seq_idx}
\title{Deduplicate a sequence index vector}
\usage{
deduplicate_seq_idx(seq_idx, hits, merge = TRUE)
}
\arguments{
\item{seq_idx}{(\code{integer}) a vector of sequence indices}

\item{hits}{(\code{data.frame}) a table of hits from, e.g.,
\code{vsearch_usearch_global()}. Should have columns \code{query} and \code{hit}, both
containing integers appearing in \code{seq_idx}.}

\item{merge}{(\code{logical}) if TRUE, merge duplicate sequence indices into a
single value}
}
\value{
a deduplicated sequence index vector; the same length as \code{seq_idx}
if \code{merge} is FALSE, otherwise (possibly) shorter.
}
\description{
Deduplicate a sequence index vector
}
