% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table.R
\name{deduplicate_seqs}
\alias{deduplicate_seqs}
\title{Remove "no-mismatch" duplicates from a sequence file}
\usage{
deduplicate_seqs(seqs, hits, outfile)
}
\arguments{
\item{seqs}{(\code{character}) file name of a sequence file}

\item{hits}{(\code{data.frame}) a table of hits from, e.g.,
\code{vsearch_usearch_global()}. Should have columns \code{query} and \code{hit}, both of
contain integer indices for sequences in \code{seqs}.}

\item{outfile}{(\code{character}) file name to write deduplicated sequences to}
}
\value{
the file name of the deduplicated sequences
}
\description{
Remove "no-mismatch" duplicates from a sequence file
}
