% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table.R
\name{deduplicate_seqtable}
\alias{deduplicate_seqtable}
\title{Remove "no-mismatch" duplicates from a mapped long sequence table}
\usage{
deduplicate_seqtable(
  seqtable,
  hits,
  abund_col = "nread",
  sample_cols = "sample",
  merge = TRUE
)
}
\arguments{
\item{seqtable}{(\code{data.frame}) a mapped long sequence table}

\item{hits}{(\code{data.frame}) a table of hits from, e.g.,
\code{vsearch_usearch_global()}. Should have columns \code{query} and \code{hit}, both of
which contain integer indices which match the \code{seq_idx} column of \code{seqtable}.}

\item{abund_col}{(\code{character}) column name containing abundance information}

\item{sample_cols}{(\code{character}) column name(s) containing sample identifiers}

\item{merge}{(\code{logical}) if TRUE, merge duplicate sequences into a single row
per sample}
}
\value{
a deduplicated sequence table
}
\description{
Remove "no-mismatch" duplicates from a mapped long sequence table
}
