% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{define_taxonomy}
\alias{define_taxonomy}
\title{Define the taxonomic ranks to use in the pipeline}
\usage{
define_taxonomy(ranks)
}
\arguments{
\item{ranks}{(\code{list} or \code{character} vector) the taxonomic ranks to use in the
pipeline, in order from most inclusive (e.g., kingdom) to least inclusive
(e.g., species). Values may be either named or unnamed. When named, the name is
taken to be the rank, and the value is the "in-group" taxon at that rank, i.e.
the taxon for which results are desired. When unnamed, the value is taken to
be the rank. Example: \code{list(kingdom = "Fungi", "phylum", "class", "order", "family", "genus", "species")}}
}
\value{
\code{NULL}.  This function is called for its side effect, which is to
configure global options.
}
\description{
Define the taxonomic ranks to use in the pipeline
}
