% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derep_map_only}
\alias{derep_map_only}
\title{Remove unneeded fields from derep object(s)}
\usage{
derep_map_only(derep)
}
\arguments{
\item{derep}{(\code{list}) \code{\link[dada2:derep-class]{derep}} object, or possibly
nested list of such objects}
}
\value{
None. This function is called for its side effect, which is to
remove the \code{uniques} and \code{quals} fields from the input object
}
\description{
Remove unneeded fields from derep object(s)
}
