% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{derepFastq}
\alias{derepFastq}
\title{Wrapper for dada2::derepFastq}
\usage{
derepFastq(fls, n = 1e+06, verbose = FALSE, qualityType = "Auto", names = fls)
}
\arguments{
\item{fls}{File names}

\item{n}{Number of reads to process}

\item{verbose}{Print progress messages}

\item{qualityType}{Quality score type}

\item{names}{Sample names for the output object}
}
\value{
A named list of 0 or more derep objects.
}
\description{
Wrapper for dada2::derepFastq
}
\details{
This function is a wrapper for \code{dada2::derepFastq} that ensures
that the output is always a list, even if it contains only one element or
none. It also allows the user to specify sample names for the output object.
}
