% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_tasks.R
\name{distribute_tasks}
\alias{distribute_tasks}
\title{Distribute tasks to bins}
\usage{
distribute_tasks(w, max_w)
}
\arguments{
\item{w}{(\code{numeric}) weight of each task}

\item{max_w}{(\code{numeric} scalar) maximum bin weight}
}
\value{
(\code{integer} vector) bin assignments
}
\description{
This function distributes tasks to bins using a greedy algorithm.
Each task is assigned to the first bin that has enough space;
or if no such bin exists, a new bin is created. Tasks with weight
greater than the maximum bin weight are assigned to a bin of their
own.
}
\examples{
w <- rev(1:8)
max_w <- 5
distribute_tasks(w, max_w)
# [1] 1 2 3 4 5 6 6 5
}
