% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{do_denovo_cluster}
\alias{do_denovo_cluster}
\title{Run denovo clustering}
\usage{
do_denovo_cluster(
  predenovo_taxon_table,
  seq_file,
  seq_file_index,
  rank,
  parent_rank,
  tax_ranks,
  denovo_thresholds,
  dist_config,
  clust_config = NULL,
  parallel_config = NULL
)
}
\arguments{
\item{predenovo_taxon_table}{(\code{data.frame}) a table of sequences for which
denovo clustering is required. Must have columns \code{seq_idx} (\code{integer})
giving the indices of sequences in \code{seq_file}, as well a column with
names matching \code{parent_rank}.}

\item{seq_file}{(\code{character}) the path to the sequence file}

\item{seq_file_index}{(\code{character} or \code{fastqindexr_index}) file path or
\code{\link[fastqindexr:create_index]{fastqindexr_index}} object for the sequence
file index}

\item{rank}{(\code{character}) the (first) taxonomic rank to calculate}

\item{parent_rank}{(\code{character}) the taxonomic rank to use as a parent}

\item{tax_ranks}{(\code{character}) the taxonomic ranks in use, from most- to
least-inclusive (e.g., from "kingdom" to "species")}

\item{denovo_thresholds}{(\code{list}) a list of thresholds to use for
denovo clustering. These are interpreted as distances on the
interval [0, 100], where 0 is a perfect match and 100 is a perfect
mismatch.}

\item{dist_config}{(\code{list}) a list of configuration options for the
distance calculation, as returned by \code{\link[optimotu:dist_config]{optimotu::dist_config()}}.}

\item{clust_config}{(\code{list}) a list of configuration options for the
clustering, as returned by \code{\link[optimotu:clust_config]{optimotu::clust_config()}}. When \code{NULL},
\code{\link[=cluster_clust_config]{cluster_clust_config()}} is evaluated (SLINK for Hamming, tree
otherwise).}

\item{parallel_config}{(\code{list}) a list of configuration options for the
parallelization, as returned by \code{\link[optimotu:parallel_config]{optimotu::parallel_config()}}. When
\code{NULL}, \code{\link[=cluster_parallel_config]{cluster_parallel_config()}} is evaluated (merge for Hamming,
concurrent otherwise).}
}
\value{
(\code{data.frame}) a table of taxonomic assignments with columns \code{seq_id}
and one column for each rank in \code{tax_ranks}; ranks from \code{parent_rank} and up
will be \code{character} vectors, while ranks from \code{rank} down will be \code{integer}
vectors (representing unnamed cluster indices)
}
\description{
This function calls \code{\link[optimotu:seq_cluster]{optimotu::seq_cluster()}} for one or more
parent taxa.
}
