% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{do_output_rdata}
\alias{do_output_rdata}
\title{Whether to write a bundled RData file of tabular outputs}
\usage{
do_output_rdata()
}
\value{
\code{logical} scalar
}
\description{
Whether to write a bundled RData file of tabular outputs
}
