% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table.R
\name{drop_from_seqtable}
\alias{drop_from_seqtable}
\title{Drop columns from a DADA2-style sequence table}
\usage{
drop_from_seqtable(seqtable, which)
}
\arguments{
\item{seqtable}{(\code{integer} matrix) a DADA2-style sequence table}

\item{which}{(\code{integer} or \code{character} vector) column indices to drop}
}
\value{
a new sequence table with the specified columns removed
}
\description{
Drop columns from a DADA2-style sequence table
}
