% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ensure_table}
\alias{ensure_table}
\title{Ensure that a table has at least one row}
\usage{
ensure_table(table)
}
\arguments{
\item{table}{(\code{data.frame}) a table to check}
}
\value{
(\code{data.frame}) the input table if it has at least one row, otherwise
a single-row table with the same columns and types, but all values \code{NA}. The
exception is if the table includes a column named "tar_group" (as produced by
\code{\link[targets:tar_group]{targets::tar_group()}}), in which case that column is given the value \code{1L} in
the output if the input had 0 rows.
}
\description{
Ensure that a table has at least one row
}
\keyword{internal}
