% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\name{epa_ng}
\alias{epa_ng}
\title{Run EPA-ng}
\usage{
epa_ng(
  ref_msa,
  tree,
  query,
  outdir = tempfile(),
  model,
  ncpu,
  epa_ng = find_epa_ng(),
  redo = TRUE,
  strip_inserts = FALSE
)
}
\arguments{
\item{ref_msa}{(\code{\link[Biostrings:XStringSet-class]{XStringSet}},
\code{\link[Biostrings:MultipleAlignment-class]{MultipleAlignment}},
or \code{character}) reference alignment, or a file name}

\item{tree}{(\code{\link[ape:read.tree]{phylo}} or \code{character}) reference tree, or
file name of a tree in Newick format}

\item{query}{(\code{\link[Biostrings:XStringSet-class]{XStringSet}},
\code{\link[Biostrings:MultipleAlignment-class]{MultipleAlignment}},
or \code{character}) query alignment, or a file name (possibly gzipped)}

\item{outdir}{(\code{character}) output directory}

\item{model}{(\code{character}) substitution model specification, or a file name
containing the model specification.  EPA-ng can parse output files from
various programs, including IQ-TREE, RAxML, and FastTree.}

\item{ncpu}{(\code{integer}) number of threads to use}

\item{epa_ng}{(\code{character}) path to the \code{epa-ng} executable}

\item{redo}{(\code{logical}) whether to overwrite existing output files}

\item{strip_inserts}{(\code{logical}) whether to strip insertions (lower-case
characters) from the query sequences before running EPA-ng}
}
\value{
(\code{character}) path to the output \code{epa_result.jplace} file. For empty
query inputs, returns a valid jplace file with no placements.
}
\description{
Run EPA-ng
}
