% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{extract_targets}
\alias{extract_targets}
\title{Extract names of targets from an expression}
\usage{
extract_targets(expr, ...)
}
\arguments{
\item{expr}{(\code{symbol}, \code{call}, or other defused expression accepted by
\code{targets::tar_deps_raw()}) an unevaluated expression containing symbols
which refer to targets tracked in the current project's
\code{targets::tar_meta()}}
}
\value{
(\code{character} vector) names of targets present in \code{expr}. If the
targets use dynamic branching so that they are stored as multiple children,
the names of the children are returned.
}
\description{
This function only works inside a running pipeline. (And maybe only on the
main thread?  TBD...)
}
\keyword{internal}
