% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fasta_combine}
\alias{fasta_combine}
\title{Interleave multiple FASTA files}
\usage{
fasta_combine(infiles, outfile, compress = FALSE)
}
\arguments{
\item{infiles}{(\code{character()}) input FASTA files, can be compressed}

\item{outfile}{(\code{character(1)}) output FASTA file}

\item{compress}{(\code{logical(1)}) compress output file}
}
\value{
\code{character(1)} output file name
}
\description{
Interleave multiple FASTA files
}
\keyword{internal}
