% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fasta_split}
\alias{fasta_split}
\title{Split a FASTA file into multiple files}
\usage{
fasta_split(infile, outfiles, compress = FALSE)
}
\arguments{
\item{infile}{Input FASTA file}

\item{outfiles}{Output FASTA files}

\item{compress}{Whether to compress the output files}
}
\value{
Output file names
}
\description{
Split a FASTA file into multiple files
}
\keyword{internal}
