% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_combine}
\alias{fastq_combine}
\title{Interleave multiple FASTQ files}
\usage{
fastq_combine(infiles, outfile, compress = FALSE)
}
\arguments{
\item{infiles}{(\code{character()}) input FASTQ files, can be compressed}

\item{outfile}{(\code{character(1)}) output FASTQ file}

\item{compress}{(\code{logical(1)}) compress output file}
}
\value{
\code{character(1)} output file name
}
\description{
Interleave multiple FASTQ files
}
\keyword{internal}
