% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_names}
\alias{fastq_names}
\title{Get the names of reads in a FASTQ file}
\usage{
fastq_names(x)
}
\arguments{
\item{x}{(\code{character}) FASTQ file path, optionally gzipped.}
}
\value{
(\code{character}) Read names.
}
\description{
Get the names of reads in a FASTQ file
}
