% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_qual_bins}
\alias{fastq_qual_bins}
\title{Get the unique quality scores in a set of FASTQ files}
\usage{
fastq_qual_bins(fastq, max_n = 1000000000L, offset = 33L)
}
\arguments{
\item{fastq}{(\code{character}) FASTQ file paths, optionally gzipped.}

\item{max_n}{(\code{integer}) Maximum number of reads to process.}

\item{offset}{(\code{integer}) ASCII offset for quality scores.}
}
\value{
(\code{integer}) Unique quality scores.
}
\description{
Get the unique quality scores in a set of FASTQ files
}
