% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_sample_multiple}
\alias{fastq_sample_multiple}
\title{Take multiple samples from a FASTQ file}
\usage{
fastq_sample_multiple(file, numerators, denominator, output, rename = FALSE)
}
\arguments{
\item{file}{(\code{character} string) the path to the input FASTQ file.}

\item{numerators}{(\code{integer} vector) a vector of numerators for the
fractions of reads to sample.}

\item{denominator}{(\code{integer} scalar) the denominator of the fraction of
reads to sample.}

\item{output}{(\code{character} vector) a vector of output file names. If an
element ends in ".gz", the output will be gzipped.}
}
\value{
a character vector of output file names
}
\description{
This function takes multiple samples from a FASTQ file using the same
denomerator and seed, but different numerators.
}
