% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_sample_number_multiple}
\alias{fastq_sample_number_multiple}
\title{Subsample a FASTQ file to multiple numbers of reads}
\usage{
fastq_sample_number_multiple(file, numbers, output, rename = FALSE)
}
\arguments{
\item{file}{(\code{character} string) the path to the input FASTQ file. May be gzipped.}

\item{numbers}{(\code{integer} vector) the numbers of reads to sample.}

\item{output}{(\code{character} vector) a vector of output file names. If an
element ends in ".gz", the output will be gzipped.}

\item{rename}{(\code{logical} scalar) whether to rename the reads in the output;
if \code{TRUE}, the read names will be replaced with a hexadecimal sequential
number.}
}
\value{
a character vector of output file names
}
\description{
This function subsamples a FASTQ file to multiple numbers of reads.
}
