% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_seq_map}
\alias{fastq_seq_map}
\title{Map the fate of individual reads through trimming and filtering steps}
\usage{
fastq_seq_map(fq_raw, fq_trim, fq_filt)
}
\arguments{
\item{fq_raw}{(\code{character}) raw FASTQ file path, optionally gzipped}

\item{fq_trim}{(\code{character}) trimmed FASTQ file path, optionally gzipped}

\item{fq_filt}{(\code{character}) filtered FASTQ file path, optionally gzipped}
}
\value{
\code{data.frame} with columns:
\itemize{
\item \code{raw_idx} (integer) index of the sequence in fastq_raw; or if sequence names
in the fastq files are hex encoded integers (e.g., during subsampling to
indicate the original index) then \code{seq_id} comes from the sequence names.
\item \code{trim_idx} (integer) index of the sequence in fastq_trim, or NA if absent
\item \code{filt_idx} (integer) index of the sequence in fastq_filt, or NA if absent
}
}
\description{
Map the fate of individual reads through trimming and filtering steps
}
