% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_split}
\alias{fastq_split}
\title{Split a FASTQ file into multiple files}
\usage{
fastq_split(infile, outfiles, compress = FALSE)
}
\arguments{
\item{infile}{Input FASTQ file}

\item{outfiles}{Output FASTQ files}

\item{compress}{Whether to compress the output files}
}
\value{
Output file names
}
\description{
Split a FASTQ file into multiple files
}
\keyword{internal}
