% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastx_split_combine.R
\name{fastx_combine}
\alias{fastx_combine}
\title{Rejoins some split fastx files}
\usage{
fastx_combine(infiles, outfile)
}
\arguments{
\item{infiles}{(\code{character(n)}) paths to the input files}

\item{outfile}{(\code{character(1)}) path to the output file}
}
\value{
(\code{character(1)}) path to the output file
}
\description{
Of course this is most useful if something happened to them in between.
If no lines are missing or reordered in the infiles, then the outfile
will end up in the same order as the original file which was split.
Otherwise this will almost certainly not happen.
}
