% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_index.R
\name{fastx_gz_hash}
\alias{fastx_gz_hash}
\title{Generate MD5 hash of a subset of sequences in a gzipped FASTA or FASTQ file}
\usage{
fastx_gz_hash(infile, index, start, n)
}
\arguments{
\item{infile}{(\code{character} filename) gzipped fasta or fastq file}

\item{index}{(\code{character} filename) index file for \code{infile}}

\item{start}{(\code{integer} scalar) one-based index to start hashing}

\item{n}{(\code{integer} scalar) number of sequences to hash}
}
\value{
(\code{character}) md5 hash
}
\description{
Generate MD5 hash of a subset of sequences in a gzipped FASTA or FASTQ file
}
