% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/zzz.R
\name{fastq_rename}
\alias{fastq_rename}
\alias{fasta_rename}
\alias{fastx_rename}
\title{Rename reads in a FAST(Q/A) file}
\usage{
fastq_rename(infile, names, outfile)

fasta_rename(infile, names, outfile)
}
\arguments{
\item{infile}{(\code{character}) FAST(A/Q) file path, optionally gzipped.}

\item{names}{(\code{character}) New read names.}

\item{outfile}{(\code{character}) FAST(A/Q) file path, optionally gzipped.}
}
\value{
(\code{character}) Output file path.
}
\description{
Rename reads in a FAST(Q/A) file
}
\section{Functions}{
\itemize{
\item \code{fastq_rename()}: Rename reads in a FASTQ file

\item \code{fasta_rename()}: Rename reads in a FASTA file

}}
