% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastx_split_combine.R
\name{fastx_split}
\alias{fastx_split}
\title{Splits a (possibly gzipped) fastx file into subfiles}
\usage{
fastx_split(infile, n, outroot = tempfile(), compress = FALSE)
}
\arguments{
\item{infile}{(\code{character(1)}) path to input file}

\item{n}{(\code{integer(1)}) number of subfiles to split into}

\item{outroot}{(\code{character(1)}) path to directory to write subfiles to}

\item{compress}{(\code{logical(1)}) whether to compress output files}
}
\value{
(\code{character(n)}) paths to the output files
}
\description{
Can handle even large files with minimal memory.
}
