% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_key.R
\name{file_to_sample_key}
\alias{file_to_sample_key}
\title{Convert a file name to a sample key}
\usage{
file_to_sample_key(filename)
}
\arguments{
\item{filename}{(\code{character}) the name of a file, possibly including path}
}
\value{
(\code{character}) the sample key
}
\description{
Convert a file name to a sample key
}
