% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\docType{data}
\name{filter_option_names}
\alias{filter_option_names}
\title{Names of options in dada2_filter_options}
\format{
An object of class \code{character} of length 2.
}
\usage{
filter_option_names
}
\description{
Names of options in dada2_filter_options
}
\keyword{datasets}
