% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{filterAndTrim}
\alias{filterAndTrim}
\title{Wrapper for dada2::filterAndTrim}
\usage{
filterAndTrim(fwd, filt, rev, filt.rev, ...)
}
\arguments{
\item{fwd}{Forward reads}

\item{filt}{Forward filtered reads}

\item{rev}{Reverse reads}

\item{filt.rev}{Reverse filtered reads}

\item{...}{Additional arguments (passed to \code{\link[dada2:filterAndTrim]{dada2::filterAndTrim()}})}
}
\value{
\code{character} vector of files names for output files
}
\description{
Wrapper for dada2::filterAndTrim
}
\details{
This function is a wrapper for \code{dada2::filterAndTrim} that ensures
that the output files are created even if they end up being empty. It also
returns an empty character vector if no input files are given.#'
}
