% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_table.R
\name{finalize_sample_table}
\alias{finalize_sample_table}
\title{Finalize a sample table after it has been read or inferred}
\usage{
finalize_sample_table(sample_table)
}
\arguments{
\item{sample_table}{(\code{tibble}) the sample table to finalize.}
}
\value{
A finalized sample table
}
\description{
This includes adding the \code{orient} column if it is not already present,
generating file names for the trimmed and filtered reads, and validation.
}
\keyword{internal}
