% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\name{find_cutadapt}
\alias{find_cutadapt}
\title{Try to find the cutadapt executable}
\usage{
find_cutadapt()
}
\value{
a \code{character} string giving the path of the cutadapt executable
}
\description{
Try to find the cutadapt executable
}
