% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\name{find_epa_ng}
\alias{find_epa_ng}
\title{Try to find the epa-ng executable}
\usage{
find_epa_ng()
}
\value{
(\code{character} string) the full path to the epa-ng executable.
}
\description{
Try to find the epa-ng executable
}
