% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\name{find_gappa}
\alias{find_gappa}
\title{Try to find the gappa executable}
\usage{
find_gappa()
}
\value{
(\code{character} string) the full path to the gappa executable.
}
\description{
Try to find the gappa executable
}
