% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{find_hmmalign}
\alias{find_hmmalign}
\title{Try to find the hmmalign executable}
\usage{
find_hmmalign()
}
\value{
a \code{character} string giving the path to the hmmalign executable
}
\description{
Try to find the hmmalign executable
}
