% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{find_hmmsearch}
\alias{find_hmmsearch}
\title{Try to find the hmmsearch executable}
\usage{
find_hmmsearch()
}
\value{
a \code{character} string giving the path to the hmmsearch executable
}
\description{
Try to find the hmmsearch executable
}
