% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{find_nhmmer}
\alias{find_nhmmer}
\title{Try to find the nhmmer executable}
\usage{
find_nhmmer()
}
\value{
a \code{character} string giving the path to the nhmmer executable
}
\description{
Try to find the nhmmer executable
}
