% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_helpers.R
\name{find_seq_col}
\alias{find_seq_col}
\title{Guess the column name in a data frame which refers to the sequence itself}
\usage{
find_seq_col(d)
}
\arguments{
\item{d}{(\code{data.frame}) data frame to search}
}
\value{
(\code{character}) name of the column
}
\description{
Guess the column name in a data frame which refers to the sequence itself
}
