% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_taxa.R
\name{find_target_taxa}
\alias{find_target_taxa}
\title{Find OTUs which contain sequences with \emph{any} probability of being a target taxon}
\usage{
find_target_taxa(target_taxa, asv_all_tax_prob, asv_taxonomy, otu_taxonomy)
}
\arguments{
\item{target_taxa}{(\code{character}) target taxa}

\item{asv_all_tax_prob}{(\code{data.frame}) ASV taxonomic probabilities, as
returned by \code{parse_protax_nameprob()} or \code{run_protax_animal()}. Should
include columns \code{seq_id}, \code{rank}, \code{taxon}, and \code{prob}.}

\item{asv_taxonomy}{(\code{data.frame}) ASV taxonomy; as \code{asv_all_tax_prob}, but
with only one taxon per rank per ASV, and \code{prob} is not required.}

\item{otu_taxonomy}{(\code{data.frame}) OTU taxonomy; as \code{asv_taxonomy}, but for OTUs.}
}
\value{
\code{data.frame} with columns \code{seq_id}, \code{asv_seq_id}, \code{otu_taxon}, \code{rank},
\code{protax_taxon}, and \code{protax_prob}.
}
\description{
Find OTUs which contain sequences with \emph{any} probability of being a target taxon
}
