% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{find_vsearch}
\alias{find_vsearch}
\title{Try to find the vsearch executable}
\usage{
find_vsearch()
}
\value{
(\code{character} string) the full path to the vsearch executable.
}
\description{
Try to find the vsearch executable
}
