% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{get_target_names}
\alias{get_target_names}
\title{Get all the target names defined in a plan}
\usage{
get_target_names(plan)
}
\arguments{
\item{plan}{(\code{targets::tar_target()} object or (possibly nested) list of such
objects) the plan to extract target names from}
}
\value{
a \code{character} vector of target names
}
\description{
Get all the target names defined in a plan
}
