% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{guild_db_target_names}
\alias{guild_db_target_names}
\title{Target names for guild database tables (and local files)}
\usage{
guild_db_target_names()
}
\value{
(\code{character}) names such as \code{guild_db_funguild} and
\code{guild_db_file_carlos}
}
\description{
Target names for guild database tables (and local files)
}
