% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_helpers.R
\name{has_alphabet}
\alias{has_alphabet}
\title{Test if all characters in a character vector are members of an alphabet.}
\usage{
has_alphabet(seq, alphabet)
}
\arguments{
\item{seq}{(\code{character}) character string(s) to test}

\item{alphabet}{(\code{character}) with all elements of width 1)}
}
\value{
(\code{logical}) \code{TRUE} if all characters in \code{seq} are also in \code{alphabet}
}
\description{
Test if all characters in a character vector are members of an alphabet.
}
\details{
This function internally uses regular expressions, so \code{alphabet}
should not begin with "^" or contain "\\".  "-", which commonly represents
a gap, is handled correctly.
}
\keyword{internal}
