% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guilds.R
\name{has_lifestyle_guild_db}
\alias{has_lifestyle_guild_db}
\title{Whether the optional lifestyle guild RDS is present}
\usage{
has_lifestyle_guild_db(path = lifestyle_guild_path())
}
\arguments{
\item{path}{(\code{character}) path to the RDS file}
}
\value{
(\code{logical}) \code{TRUE} if \code{path} exists
}
\description{
Whether the optional lifestyle guild RDS is present
}
