% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_helpers.R
\name{hash_sequences}
\alias{hash_sequences}
\alias{hash_sequences.character}
\alias{hash_sequences.XStringSet}
\alias{hash_sequences.data.frame}
\title{Hash sequences}
\usage{
hash_sequences(seq, use_names = TRUE, ...)

\method{hash_sequences}{character}(seq, use_names = TRUE, ...)

\method{hash_sequences}{XStringSet}(seq, use_names = TRUE, ...)

\method{hash_sequences}{data.frame}(
  seq,
  use_names = TRUE,
  seq_col = find_seq_col(seq),
  name_col = if (isTRUE(use_names)) find_name_col(seq) else NULL,
  ...
)
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character} string representing a file name,
named \code{character} vector, or \code{Biostrings::XStringSet}) sequences to hash}

\item{use_names}{(\code{logical}) if \code{TRUE}, the result hashes will be named using
the names of the sequences.}

\item{...}{additional arguments to pass to the hashing function}

\item{seq_col}{(\code{character} string) name of the column in \code{seq} which
contains the sequences}

\item{name_col}{(\code{character} string) name of the column in \code{seq} which
contains the sequence names}
}
\value{
(\code{integer} vector) hash codes
}
\description{
Hash sequences
}
