% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_table.R
\name{infer_sample_table}
\alias{infer_sample_table}
\title{Infer sample table from raw sequence file names}
\usage{
infer_sample_table(
  raw_path = "sequences/01_raw",
  file_extension = read_file_extension()
)
}
\arguments{
\item{raw_path}{(\code{character}) the path to the directory containing the raw
sequence files.}

\item{file_extension}{(\code{character}) a regular expression that matches the
file extension of the sequence files.}
}
\value{
A \code{tibble} with columns \code{fastq_R1}, \code{fastq_R2}, \code{seqrun}, \code{sample},
\code{neg_control}, and \code{pos_control}. The \code{fastq_R1} and \code{fastq_R2} columns
contain the paths to the forward and reverse read files, respectively.
The \code{seqrun} column contains the name of the sequencing run (the directory
containing the read files), and the \code{sample} column contains the sample
name extracted from the file names. The \code{neg_control} and \code{pos_control}
columns are logical vectors indicating whether the sample is a negative or
positive control, respectively.
}
\description{
This function infers a sample table from the directory structure and file
names of paired FASTQ files.
}
\details{
This function is designed for the case where:
\enumerate{
\item Each sequencing run is in a separate subdirectory of \code{raw_path}. The
actual read files can be nested in further subdirectories below this.
\item The read files are named according to the Illumina convention, with
\code{R1} and \code{R2} suffixes for the forward and reverse reads, respectively.
\item The first part of the file name is the sample name.
\item Between the sample name and the \code{R1}/\code{R2} suffix, there may be sample and
lane indices of the form "S{nnn}_L{nnn}".  These are ignored.
\item Sample name, sample/lane indices, and \code{R1}/\code{R2} suffixes may be separated
by underscores or dots.
\item After the \code{R1}/\code{R2} suffix, there may be an additional \verb{_nnn} suffix,
which is ignored.
\item The file extension should match the regular expression in \code{file_extension}.
The default value matches \code{.fastq}, \code{.fq}, \code{.fastq.gz}, and \code{.fq.gz}, which
should cover the majority of cases.
\item Sample named including the test "NEGPCR", "NEGEXT", "NEGATIVE",
"NEGCONTROL", "NEG_CONTROL", or "BLANK" (matched case-insensitively) are
assumed to be negative controls, and those including "MOCK", "AMPTK",
"POSITIVE", "POSCONTROL", or "POS_CONTROL" are assumed to be positive
controls.
}
}
