% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\name{is_jplace}
\alias{is_jplace}
\title{Test whether a jplace object is valid}
\usage{
is_jplace(jplace)
}
\arguments{
\item{jplace}{(\code{list}) a jplace object}
}
\value{
(\code{logical}) whether the jplace object is valid
}
\description{
Test whether a jplace object is valid
}
\keyword{internal}
