% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_local}
\alias{is_local}
\title{Determine if the R process seems to be running "locally" (i.e. not on a cluster)}
\usage{
is_local()
}
\value{
a logical value indicating whether the R process is running locally
}
\description{
Determine if the R process seems to be running "locally" (i.e. not on a cluster)
}
