% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_slurm}
\alias{is_slurm}
\title{Determine if the R process seems to be running inside a SLURM job}
\usage{
is_slurm()
}
\value{
a logical value indicating whether the R process is running inside a
SLURM job
}
\description{
Determine if the R process seems to be running inside a SLURM job
}
