% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_snakemake}
\alias{is_snakemake}
\title{Determine if the R process seems to be running inside a Snakemake job}
\usage{
is_snakemake()
}
\value{
a logical value indicating whether the R process is running inside a
Snakemake job
}
\description{
Determine if the R process seems to be running inside a Snakemake job
}
