% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{join_table_with_taxsort}
\alias{join_table_with_taxsort}
\title{Join a taxon table with \code{asv_taxsort}}
\usage{
join_table_with_taxsort(taxon_table, asv_taxsort)
}
\arguments{
\item{taxon_table}{(\code{data.frame}) known or closed-reference taxon table.}

\item{asv_taxsort}{(\code{data.frame}) ASV taxonomic sort table.}
}
\value{
(\code{data.frame}) joined table.
}
\description{
Join a taxon table with \code{asv_taxsort}
}
\keyword{internal}
