% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{known_ranks}
\alias{known_ranks}
\alias{set_known_ranks}
\title{Get or set the taxonomic ranks which are assumed to be "known"}
\usage{
known_ranks()

set_known_ranks(value)
}
\arguments{
\item{value}{(\code{character} vector) the taxonomic ranks to assume are "known"}
}
\value{
a \code{character} vector of taxonomic ranks

the previously set known ranks
}
\description{
Get or set the taxonomic ranks which are assumed to be "known"
}
