% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{known_taxa}
\alias{known_taxa}
\alias{set_known_taxa}
\title{Get or set the taxa which are assumed to be "known"}
\usage{
known_taxa()

set_known_taxa(value)
}
\arguments{
\item{value}{(\code{character} vector) the taxa to assume are "known"}
}
\value{
a \code{character} vector of taxa, with the same length as \code{known_ranks()}

the previously set known taxa
}
\description{
Get or set the taxa which are assumed to be "known"
}
