% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krona.R
\name{krona_html_attributes}
\alias{krona_html_attributes}
\title{Krona HTML attribute and color header}
\usage{
krona_html_attributes(ranks = tax_ranks())
}
\arguments{
\item{ranks}{(\code{character}) taxonomic ranks, most inclusive first}
}
\value{
(\code{character}) XML lines
}
\description{
XML lines for the \verb{<attributes>} block plus the default color mapping
(tip-most unknown rank).
}
