% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krona.R
\name{krona_xml_nodes}
\alias{krona_xml_nodes}
\title{Write a KronaTools file}
\usage{
krona_xml_nodes(
  data,
  .rank,
  maxrank = rank2factor(tip_rank()),
  outfile,
  pre = NULL,
  post = NULL,
  taxonomy = paste(known_taxa(), collapse = ","),
  node_data_format = NULL,
  node_xml_format = xml_format(node_data_format),
  ...
)
}
\arguments{
\item{data}{(\code{data.frame}) as generated by \code{generate_krona_data()}}

\item{.rank}{(\code{character} or \code{factor}) the rank to generate nodes for}

\item{maxrank}{(\code{factor}) the maximum rank to generate nodes for}

\item{outfile}{(\code{character} or \code{connection}) the file to write the XML to}

\item{pre}{(\code{character}) text to write before the XML nodes}

\item{post}{(\code{character}) text to write after the XML nodes}

\item{taxonomy}{(\code{character}) the taxonomic prefix to filter the data by}

\item{node_data_format}{(\code{list}) a named list whose elements are values to be
formatted as data for a KronaTools XML file}

\item{node_xml_format}{(\code{character}) the XML format to use for each node}

\item{...}{additional arguments; passed during recursion but currently ignored}
}
\value{
(\code{character}) the name of the file written
}
\description{
Write a KronaTools file
}
