% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{learnErrors}
\alias{learnErrors}
\title{Wrapper for dada2::learnErrors}
\usage{
learnErrors(
  fls,
  nbases = 1e+09,
  nreads = NULL,
  errorEstimationFunction = choose_dada_error_function,
  multithread = FALSE,
  randomize = FALSE,
  MAX_CONSIST = 10,
  OMEGA_C = 0,
  qualityType = "Auto",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{fls}{File names}

\item{nbases}{Number of bases to use for error estimation}

\item{nreads}{Number of reads to use for error estimation}

\item{errorEstimationFunction}{Function to estimate error rates}

\item{multithread}{Use multiple threads}

\item{randomize}{Randomize the order of reads}

\item{MAX_CONSIST}{Maximum number of consistent bases}

\item{OMEGA_C}{Omega constant}

\item{qualityType}{Quality score type}

\item{verbose}{Print progress messages}

\item{...}{Additional arguments (passed to \code{\link[dada2:learnErrors]{dada2::learnErrors()}})}
}
\value{
An error object or \code{NULL} if no input files are given
}
\description{
Wrapper for dada2::learnErrors
}
