% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guilds.R
\name{lifestyle_guild_path}
\alias{lifestyle_guild_path}
\title{Default path to the Carlos/lifestyle guild table}
\usage{
lifestyle_guild_path()
}
\value{
(\code{character}) relative path used by the pipeline
}
\description{
Default path to the Carlos/lifestyle guild table
}
