% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guilds.R
\name{load_guild_database}
\alias{load_guild_database}
\title{Load a guild database specified in \code{pipeline_options.yaml}}
\usage{
load_guild_database(source, path = NULL)
}
\arguments{
\item{source}{(\code{character}) one of \code{download}, \code{lifestyle}, or \code{file}}

\item{path}{(\code{character}) file path; ignored when \code{source} is \code{download}}
}
\value{
\code{data.frame} suitable for \code{FUNGuildR::funguild_assign()}
}
\description{
Load a guild database specified in \code{pipeline_options.yaml}
}
