% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{local_cpus}
\alias{local_cpus}
\title{Get the number of CPUs available to the R process}
\usage{
local_cpus()
}
\value{
an integer indicating the number of CPUs available to the R process
}
\description{
If running inside a Snakemake or SLURM job, this function will return the
number of CPUs allocated to the job. Otherwise, it will return the value of
the option \code{"optimotu_num_threads"} if it is set, otherwise the number of
CPUs available to the R process minus one.
}
