% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulu_long.R
\name{lulu_distmx}
\alias{lulu_distmx}
\title{Compute pairwise distances for LULU}
\usage{
lulu_distmx(
  seqall_file,
  seqtable,
  seqall_index = NULL,
  threshold = lulu_max_dist(),
  dist_config = lulu_dist_config(),
  parallel_config = optimotu::parallel_concurrent(local_cpus()),
  ...
)
}
\arguments{
\item{seqall_file}{(\code{character} string) name(s) of one or more gzipped FASTA
files containing the master list of sequences. Headers should be parsable as
integers, and should be the 1-based index of the sequence in the file.}

\item{seqtable}{(\code{data.frame}) the long-format table to denoise. Must have
columns \code{seq_idx} (\code{integer}) and \code{nread} (\code{integer}), where \code{seq_idx} refers
to the index of sequences in \code{seqall_file}. Additional columns
are ignored.}

\item{seqall_index}{(\code{character} or \code{fastqindexr_index}, or \code{NULL}) file path
or \code{\link[fastqindexr:create_index]{fastqindexr_index}} object for
\code{seqall_file}, as from \code{\link[=write_fastqindexr_index]{write_fastqindexr_index()}}, \code{\link[=fastx_gz_index]{fastx_gz_index()}}, or
\code{\link[fastqindexr:create_index]{fastqindexr::create_index()}}. Character paths may be \code{.fqi} or \code{.qs2}.
Not supported for multiple input files; in that case the files will be
fully read into memory one at a time.}

\item{threshold}{(\code{numeric} scalar) maximum pairwise distance to report
between two sequences.}

\item{dist_config}{(\code{optimotu_dist_config}) configuration for distance
calculation, as returned by \code{\link[optimotu:dist_config]{optimotu::dist_config()}} or its helpers.}

\item{parallel_config}{(\code{optimotu_parallel_config}) configuration for
parallel execution, as returned by \code{\link[optimotu:parallel_config]{optimotu::parallel_config()}} or its
helpers. Only the number of threads is used.}

\item{...}{additional arguments for dependency tracking. These are ignored.}
}
\value{
a \code{data.frame} with columns \code{seq_idx1}, \code{seq_idx2}, \code{dist},
\code{align_length}, \code{n_gap}, \code{max_gap}, \code{nread1}, and \code{nread2}.
}
\description{
This is a pipeline step which calculates pairwise distances between
sequences in a single sample. It is intended to be used inside
\code{dplyr::summarize()} after grouping by sample.
}
\keyword{internal}
