% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulu_long.R
\name{lulu_map_combine}
\alias{lulu_map_combine}
\title{Combine sparse scoped LULU maps into a full OTU map}
\usage{
lulu_map_combine(stats, ...)
}
\arguments{
\item{stats}{(\code{data.frame} or target expression) output of
\code{\link[=lulu_otu_stats]{lulu_otu_stats()}}}

\item{...}{sparse map \code{data.frame}s, or expressions containing map targets
(as with \code{\link[=lulu_map_lowmem]{lulu_map_lowmem()}}). Each map must have columns \code{seq_idx} and
\code{lulu_idx}.}
}
\value{
a full \code{data.frame} with \code{seq_idx} and path-compressed \code{lulu_idx}
}
\description{
Combine sparse scoped LULU maps into a full OTU map
}
