% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lulu_map_impl}
\alias{lulu_map_impl}
\title{LULU secondary denoising}
\usage{
lulu_map_impl(
  match_id1,
  match_id2,
  match_nread1,
  match_nread2,
  match_dist,
  seq_idx,
  nread,
  max_dist,
  min_abundance_ratio = 1,
  min_cooccurrence_ratio = 0.95,
  use_mean_abundance_ratio = FALSE,
  verbose = 0L
)
}
\arguments{
\item{match_id1}{(\code{integer}) vector of sequence IDs for the first sequence
in each match; sequence IDs should be assigned such that when the sequences
are in ID order, they are sorted by decreasing occurrence, with ties broken
by decreasing abundance.}

\item{match_id2}{(\code{integer}) vector of sequence IDs for the second sequence
in each match}

\item{match_nread1}{(\code{integer}) vector of read counts for the first sequence
in each match}

\item{match_nread2}{(\code{integer}) vector of read counts for the second
sequence in each match}

\item{match_dist}{(\code{numeric}) vector of distances between the two sequences
in each match}

\item{seq_idx}{(\code{integer}) vector of sequence IDs. Each ID should occur
once per sample where the sequence occurs.}

\item{nread}{(\code{integer}) vector of read counts for each sequence in each
sample}

\item{max_dist}{(\code{numeric}) maximum distance for two sequences to be
considered as parent-child}

\item{min_abundance_ratio}{(\code{numeric}) minimum abundance ratio for two
sequences to be considered as parent-child}

\item{min_cooccurrence_ratio}{(\code{numeric}) minimum co-occurrence ratio for
two sequences to be considered}

\item{use_mean_abundance_ratio}{(\code{logical}) if \code{TRUE}, \code{min_abundance_ratio}
is interpreted as a minimum value for the mean of the abundance ratios in
all samples where the two sequences co-occur.  Otherwise (default) it is
interpreted as a minimum value for the the abundance ratio of all samples
where the two sequences co-occur.}

\item{verbose}{(\code{integer}) level of verbosity. At level 0 (default),
no messages are printed.}
}
\value{
a two-column \code{data.frame} with columns \code{seq_idx} and \code{lulu_idx}.
\code{seq_idx} includes all values which occur in the \code{seq_idx} argument, and
\code{lulu_idx} gives the index of the denoised sequence.
}
\description{
The "match" inputs are intended to be calculated by pairwise distance
calculations within each sample, the results of which are concatenated
together. This requires some duplicate calculations for pairs that occur in
many samples, but prevents the need to calculate distances for pairs that
never co-occur, which is a large number of pairs.
}
