% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulu_long.R
\name{lulu_map_lowmem}
\alias{lulu_map_lowmem}
\title{LULU secondary denoising for big(ger) data}
\usage{
lulu_map_lowmem(
  otu_table,
  match_table,
  max_dist = lulu_max_dist(),
  min_abundance_ratio = lulu_min_abundance_ratio(),
  min_cooccurrence_ratio = lulu_min_cooccurrence_ratio(),
  use_mean_abundance_ratio = lulu_use_mean_abundance_ratio(),
  id_is_int = TRUE,
  id_is_sorted = FALSE,
  verbose = 0
)
}
\arguments{
\item{otu_table}{(expression containing one or more \code{targets} targets which
evaluate to \code{data.frame}) the long-format table to denoise. Must have
columns \code{seq_idx} (\code{integer}) and \code{nread} (\code{integer}).  Additional columns
such as the sample id or metadata are allowed but ignored.}

\item{match_table}{(expression containing one or more \code{targets} targets which
evaluate to \code{data.frame}) table of pairwise matches between co-occuring OTUs.
Must have columns \code{seq_idx1} and \code{seq_idx2} (both \code{integer}), as well as
\code{nread1} and \code{nread2} (\code{integer}) and \code{dist} (\code{numeric}). Additional columns
are ignored.}

\item{max_dist}{(\code{numeric} scalar) maximum pairwise distance for two OTUs to
be considered for merging. This may be given as a percentage between 0 and
100, as a fraction between 0.0 and 1.0, or any other non-negative scale, but
should be on the same scale as the values of \code{dist} in \code{match_table}.
Default: \code{0.1} (intended as a fraction).}

\item{min_abundance_ratio}{(\code{numeric} scalar) minimum ratio of
"parent":"daughter" abundances for two OTUs to be considered for merging.
The exact meaning is dependent on the value of \code{use_mean_abundance_ratio}.
Default: \code{1}.}

\item{min_cooccurrence_ratio}{(\code{numeric} scalar) minimum ratio of
co-occurrences to total occurrences of the "daughter" for two OTUs to be
considered for merging. Default: \code{1}.}

\item{use_mean_abundance_ratio}{(\code{logical} flag) if \code{TRUE},
\code{min_abundance_ratio} refers to the mean of the relative abundances between
the "parent" and "daughter" sequences in all samples where they co-occur. If
\code{FALSE}, the minimum is enforced for every sample. Default: \code{FALSE}.}

\item{id_is_int}{(\code{logical} flag) if \code{TRUE}, the OTU identifiers are
integers; in this case they are expected to be in columns \verb{seq_idx*}. This
version does not yet support \code{FALSE} (character identifiers); see Details.}

\item{id_is_sorted}{(\code{logical} flag) if \code{TRUE}, the values of \verb{seq_idx*}
are sorted in decreasing order of prevalence (number of samples), with ties
broken by decreasing total abundance (\code{nread}). This version does not yet
support \code{TRUE}; see Details.}

\item{verbose}{(\code{integer} scalar) level of verbosity for progress messages.}
}
\value{
a two-column \code{data.frame} with columns \code{seq_idx} and \code{lulu_idx}
or \code{seq_id} and \code{lulu_id}. \verb{seq_id*} includes all values which occur in
\code{otu_table}, and \verb{lulu_id*} gives the corresponding value in \code{lulu_table};
these differ only for sequences which were determined to be minor variants
of another sequence.
}
\description{
This is an alternative implementation of \code{lulu_map()} for integration in a
\code{targets} pipeline. It uses non-standard evaluation to handle the two table
arguments (\code{otu_table} and \code{match_table}). It should be run in a target with
option \code{retrieval = "none"}. The assumption is that the LULU match table
and/or the OTU occurrence table are split into multiple files on disk using
dynamic and/or static branching in targets, and that it may be too memory
intensive to simply load all of the files to generate the full tables in
memory. Instead, they are processed one at a time, with calls to \code{gc()}
after each one. This saves memory in two ways: first, it avoids holding the
full data in both its R representation and in its internal C++ representation
simultaneously; second, the internal C++ structures are smaller
than the R tables.
}
\details{
The actual expression passed to \code{otu_table} and \code{match_table} is not
evaluated. Instead the expression is "defused" and all of the symbols it
contains are checked against the list of targets in the current pipeline.
Any that match are loaded individually for processing. The final result is
the same as if the input targets had been concatenated using \code{rbind()} and
passed to \code{lulu_map()}.  This means you should NOT pass any expression which
is intended to modify the inputs, especially one which contains other targets
from the pipeline which are not (pieces of) the OTU table or LULU match
table.

Because the function should be run in a target with \code{retrieval = "none"},
any other arguments which depend on other targets should have those targets
manually loaded. Do not call \code{targets::tar_read()} from a crew worker;
that is forbidden during a running pipeline. Prefer splicing extra
values into the command with \verb{!!} at plan time.

This version does not have the full functionality of \code{lulu_map()}; in
particular it gives an error id \code{id_is_int = FALSE} and a warning if
\code{id_is_sorted = TRUE} (it always behaves as if \code{id_is_sorted = FALSE}, but
this does not actually cause a problem if the IDs are in fact sorted).
}
