% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lulu_map_lowmem_impl}
\alias{lulu_map_lowmem_impl}
\title{LULU secondary denoising for "big" data targets pipeline}
\usage{
lulu_map_lowmem_impl(
  otu_table_names,
  match_table_names,
  max_dist,
  min_abundance_ratio = 1,
  min_cooccurrence_ratio = 0.95,
  use_mean_abundance_ratio = FALSE,
  verbose = 0L
)
}
\arguments{
\item{otu_table_names}{(\code{character} vector) fully resolved target names for
the OTU occurrence table.}

\item{match_table_names}{(\code{character} vector) fully resolved target names
for the lulu match table, as generated by \code{lulu_distmx()}.}

\item{max_dist}{(\code{numeric}) maximum distance for two sequences to be
considered as parent-child}

\item{min_abundance_ratio}{(\code{numeric}) minimum abundance ratio for two
sequences to be considered as parent-child}

\item{min_cooccurrence_ratio}{(\code{numeric}) minimum co-occurrence ratio for
two sequences to be considered}

\item{use_mean_abundance_ratio}{(\code{logical}) if \code{TRUE}, \code{min_abundance_ratio}
is interpreted as a minimum value for the mean of the abundance ratios in
all samples where the two sequences co-occur.  Otherwise (default) it is
interpreted as a minimum value for the the abundance ratio of all samples
where the two sequences co-occur.}

\item{verbose}{(\code{integer}) level of verbosity. At level 0 (default),
no messages are printed.}
}
\description{
This version of \code{lulu_map()} is intended for integration in a \code{targets}
pipeline, where it should be used inside a target with option
\code{retrieval = "none"}. The assumption is that the LULU match table and/or the
OTU occurrence table are split into multiple files on disk using
dynamic and/or static branching in targets, and that it may be too memory
intensive to simply load all of the files to generate the full tables in
memory. Instead, they are processed one at a time, with calls to \code{gc()}
after each one. This saves memory in two ways: first, it avoids holding the
full data in both its R representation and in its internal C++
simultaneously; second, the internal C++ structures are smaller
than the R tables.
}
