% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lulu_otu_stats_dfs_impl}
\alias{lulu_otu_stats_dfs_impl}
\title{In-memory variant of \code{\link[=lulu_otu_stats_impl]{lulu_otu_stats_impl()}} for tests}
\usage{
lulu_otu_stats_dfs_impl(otu_tables, seqrun_ids, verbose = 0L)
}
\arguments{
\item{otu_tables}{(\code{list}) of OTU \code{data.frame}s with \code{seq_idx} and \code{nread}.}

\item{seqrun_ids}{(\code{integer}) parallel to \code{otu_tables}.}

\item{verbose}{(\code{integer}) verbosity level.}
}
\value{
same structure as \code{\link[=lulu_otu_stats_impl]{lulu_otu_stats_impl()}}.
}
\description{
In-memory variant of \code{\link[=lulu_otu_stats_impl]{lulu_otu_stats_impl()}} for tests
}
