% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lulu_long.R
\name{lulu_table}
\alias{lulu_table}
\title{Apply LULU denoising to an OTU table}
\usage{
lulu_table(lulu_map, otu_table)
}
\arguments{
\item{lulu_map}{(\code{data.frame}) output of a call to \code{lulu_map()}}

\item{otu_table}{(\code{data.frame}) the long-format table to denoise. Must have
columns \code{seq_id} (\code{character}) xor \code{seq_idx} (\code{integer}), as well as
\code{nread} (\code{integer}). Additional columns should collectively define the
sample. Typically this includes the sample name, and sample-level metadata
are also fine to include, but sequence-level data are not. If no such values
are included, then the entire table is treated as a single sample.}
}
\value{
a \code{data.frame} with the same columns as \code{otu_table}, but with rows
combined according to \code{lulu_map}.
}
\description{
Apply LULU denoising to an OTU table
}
