% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table.R
\name{make_long_sequence_table}
\alias{make_long_sequence_table}
\alias{make_long_sequence_table.data.frame}
\alias{make_long_sequence_table.matrix}
\alias{make_long_sequence_table.list}
\title{Convert an object into a long (i.e. sparse) sequence occurrence table}
\usage{
make_long_sequence_table(x, rc = FALSE)

\method{make_long_sequence_table}{data.frame}(x, rc = FALSE)

\method{make_long_sequence_table}{matrix}(x, rc = FALSE)

\method{make_long_sequence_table}{list}(x, rc = FALSE)
}
\arguments{
\item{x}{(\code{data.frame} as returned by \code{dada2::mergePairs()}, or integer matrix
as returned by \code{dada2::makeSequenceTable()}, or a list of one of these.)}

\item{rc}{(logical flag) if TRUE, sequences in \code{x} will be reverse complemented.}
}
\value{
a \code{data.frame} with columns \code{sample}, \code{seq}, and \code{nread}
}
\description{
Convert an object into a long (i.e. sparse) sequence occurrence table
}
