% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table.R
\name{make_mapped_sequence_table}
\alias{make_mapped_sequence_table}
\alias{make_mapped_sequence_table.data.frame}
\alias{make_mapped_sequence_table.matrix}
\alias{make_mapped_sequence_table.uc_cluster}
\alias{make_mapped_sequence_table.list}
\title{Convert an object into a long (i.e. sparse) sequence occurrence table where
sequences are stored as integer indices to a master list}
\usage{
make_mapped_sequence_table(x, seqs, rc = FALSE)

\method{make_mapped_sequence_table}{data.frame}(x, seqs, rc = FALSE)

\method{make_mapped_sequence_table}{matrix}(x, seqs, rc = FALSE)

\method{make_mapped_sequence_table}{uc_cluster}(x, seqs, rc = FALSE)

\method{make_mapped_sequence_table}{list}(x, seqs, rc = FALSE)
}
\arguments{
\item{x}{(\code{data.frame} as returned by \code{dada2::mergePairs()}, integer
matrix as returned by \code{dada2::makeSequenceTable()}, \code{uc_cluster} as
returned by \code{\link[=vsearch_cluster_unoise2]{vsearch_cluster_unoise2()}}, or a list of one of these.)}

\item{seqs}{(\code{character} vector, file name, or \code{Biostrings::XStringSet})
master list of sequences}

\item{rc}{(logical flag) if TRUE, sequences in \code{x} will be reverse
complemented.}
}
\value{
a \code{data.frame} with columns \code{sample}, \code{seq_idx}, and \code{nread}
}
\description{
Convert an object into a long (i.e. sparse) sequence occurrence table where
sequences are stored as integer indices to a master list
}
