% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_helpers.R
\name{make_seq_names}
\alias{make_seq_names}
\title{Generate names for a set of sequences}
\usage{
make_seq_names(n, prefix)
}
\arguments{
\item{n}{(\code{integer}) number of sequences to name}

\item{prefix}{(\code{character}) prefix to use for the names}
}
\value{
(\code{character} vector) names
}
\description{
Names produced are like "ASV0001", "ASV0002", ... The prefix ("ASV" in
the example) is customizable, and the number will be zero-padded to fit
all values.
}
