% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{makeBinnedQualErrfun}
\alias{makeBinnedQualErrfun}
\title{Error function for binned quality scores}
\usage{
makeBinnedQualErrfun(binnedQ = c(2, 11, 25, 37))
}
\arguments{
\item{binnedQ}{Vector of binned quality scores}
}
\value{
A function that estimates error rates based on binned quality scores
}
\description{
Error function for binned quality scores
}
\details{
From dada2 commit \url{https://github.com/benjjneb/dada2/commit/7714487b153ca133cb6f03cb01d09fc05be60159}
}
