% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_map.R
\name{merge_read_maps}
\alias{merge_read_maps}
\title{Merge forward and reverse read maps}
\usage{
merge_read_maps(read_map_fwd, read_map_rev)
}
\arguments{
\item{read_map_fwd}{(\code{data.frame}) forward read map, as returned by
\code{\link[=dada2_read_map]{dada2_read_map()}} or \code{\link[=unoise_read_map]{unoise_read_map()}}}

\item{read_map_rev}{(\code{data.frame}) reverse read map, as returned by
\code{\link[=dada2_read_map]{dada2_read_map()}} or \code{\link[=unoise_read_map]{unoise_read_map()}}}
}
\description{
Note that this is only needed for workflows where the sequences are not all
in the same orientation, not for ordinary Illumina R1 and R2. (Which are
merged earlier using \code{dada2::mergePairs()} or vsearch mergepairs.)
}
