% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_map.R
\name{merge_seq_maps}
\alias{merge_seq_maps}
\title{Merge forward and reverse sequence maps}
\usage{
merge_seq_maps(seqmap_fwd, seqmap_rev)
}
\arguments{
\item{seqmap_fwd}{(\code{data.frame}) forward sequence map, as returned by \code{seq_map()}}

\item{seqmap_rev}{(\code{data.frame}) reverse sequence map, as returned by \code{seq_map()}}
}
\description{
Note that this is only needed for workflows where the sequences are not all in
the same orientation, not for ordinary Illumina R1 and R2. (Which are merged
earlier using \code{dada2::mergePairs()}.)
}
