% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{merged_filter_option_names}
\alias{merged_filter_option_names}
\title{Names of options in merged_filter_options}
\usage{
merged_filter_option_names
}
\description{
Names of options in merged_filter_options
}
