% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{mergePairs}
\alias{mergePairs}
\title{Wrapper for dada2::mergePairs}
\usage{
mergePairs(
  dadaF,
  derepF,
  dadaR,
  derepR,
  minOverlap = 12,
  maxMismatch = 0,
  returnrejects = FALSE,
  propagateCol = character(0),
  justConcatenate = FALSE,
  trimOverhang = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dadaF}{Forward denoised reads}

\item{derepF}{Forward dereplicated reads}

\item{dadaR}{Reverse denoised reads}

\item{derepR}{Reverse dereplicated reads}

\item{minOverlap}{Minimum overlap}

\item{maxMismatch}{Maximum mismatch}

\item{returnrejects}{Return rejected reads}

\item{propagateCol}{Columns to propagate}

\item{justConcatenate}{Just concatenate reads}

\item{trimOverhang}{Trim overhang}

\item{verbose}{Print progress messages}

\item{...}{Additional arguments (passed to \code{\link[dada2:mergePairs]{dada2::mergePairs()}})}
}
\value{
A \code{data.frame} as returned by \code{\link[dada2:mergePairs]{dada2::mergePairs()}}, or a \code{list} of
such \code{data.frame}s.  In partticular, if the inputs \code{dadaF}, \code{derepF}, \code{dadaR},
and \code{derepR} are all lists, even of length 0 or 1, the output will also be a
list of the same length.
}
\description{
Wrapper for dada2::mergePairs
}
