% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_helpers.R
\name{name_seqs}
\alias{name_seqs}
\alias{name_seqs.XStringSet}
\alias{name_seqs.character}
\alias{name_seqs.data.frame}
\alias{name_seqs.matrix}
\title{Name a set of sequences}
\usage{
name_seqs(seq, prefix, ...)

\method{name_seqs}{XStringSet}(seq, prefix, ...)

\method{name_seqs}{character}(seq, prefix, ...)

\method{name_seqs}{data.frame}(seq, prefix, id_col = prefix, ...)

\method{name_seqs}{matrix}(seq, prefix, ...)
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character}, or \code{Biostrings::XStringSet}) sequences
to name}

\item{prefix}{(\code{character}) prefix to use for the names}

\item{...}{additional arguments to pass to methods}

\item{id_col}{(\code{character} string) name of the column in \code{seq} which
contains the sequence IDs}
}
\value{
(same type of object as \code{seq}) named sequences
}
\description{
Sequences are named using \code{make_seq_names()}.
}
\details{
If the input is a file name, the file will be modified.
}
