% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{nomismatch_hits_vsearch}
\alias{nomismatch_hits_vsearch}
\title{Collapse ASVs which are identical in their overlaps using VSEARCH}
\usage{
nomismatch_hits_vsearch(
  seqtab,
  seqs = NULL,
  abund_col = "nread",
  fastx_index = NULL,
  ...,
  ncpu = local_cpus()
)
}
\arguments{
\item{seqtab}{(\code{data.frame}) long sequence table as returned by
\code{make_long_sequence_table()} or \code{make_mapped_sequence_table()}}

\item{seqs}{(\code{NULL}, \code{data.frame},
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, \code{character} vector, or
file name) if seqtab is a mapped sequence table, then the master sequence
list, otherwise should be \code{NULL}.}

\item{abund_col}{(\code{character} string) name of the column in \code{seqtab} which
contains the abundance of each sequence}

\item{fastx_index}{(\code{NULL} of file name) if \code{seqs} is a file name pointing to
a gzipped FASTA file, then the (optional) index file for that file}

\item{...}{additional arguments to pass to \code{vsearch_cluster_smallmem()}}

\item{ncpu}{(\code{integer} count) number of threads to use}
}
\value{
\code{data.frame} with columns \code{seq_idx_in} and \code{seq_idx_out}, where
\code{seq_idx_in} is the index of a sequence in \code{seqtab}, and \code{seq_idx_out} is
the index of the sequence which is the centroid of the cluster containing
\code{seq_idx_in}
}
\description{
Collapse ASVs which are identical in their overlaps using VSEARCH
}
