% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{output_table_formats}
\alias{output_table_formats}
\title{Per-file tabular output formats (excludes \code{rdata})}
\usage{
output_table_formats()
}
\value{
\code{character} vector of format names for \code{\link[=write_tabular_outputs]{write_tabular_outputs()}}
}
\description{
Per-file tabular output formats (excludes \code{rdata})
}
