% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{pack_precluster_taxon_table}
\alias{pack_precluster_taxon_table}
\title{Pack parent taxa into \code{tar_group} bins of at most \code{max_ops}}
\usage{
pack_precluster_taxon_table(table, parent_rank, max_ops)
}
\arguments{
\item{table}{(\code{data.frame}) precluster table with \code{ops} and \code{tar_group}.}

\item{parent_rank}{(\code{character}) parent taxon column name.}

\item{max_ops}{(\code{numeric} scalar) maximum ops per execution group.}
}
\value{
(\code{data.frame}) table with \code{tar_group} from \code{\link[=distribute_tasks]{distribute_tasks()}}.
}
\description{
Pack parent taxa into \code{tar_group} bins of at most \code{max_ops}
}
\keyword{internal}
