% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\name{parse_gappa_per_query}
\alias{parse_gappa_per_query}
\title{Parse Gappa per-query results}
\usage{
parse_gappa_per_query(per_query, ranks, id_is_int = FALSE)
}
\arguments{
\item{per_query}{(\code{character}) path to the per-query results file}

\item{ranks}{(\code{character}) vector of taxonomic ranks}

\item{id_is_int}{(\code{logical}) whether the sequence IDs are integers}
}
\description{
Parse Gappa per-query results
}
\keyword{internal}
