% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\name{parse_iqtree_log}
\alias{parse_iqtree_log}
\title{Parse IQ-TREE model specification}
\usage{
parse_iqtree_log(model)
}
\arguments{
\item{model}{(\code{character}) model specification or a file name}
}
\description{
Parse IQ-TREE model specification
}
\details{
Parses IQ-TREE output into an EPA-ng compatible model string including fitted
substitution rates, frequencies, and heterogeneity parameters when available.
}
