% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{parse_output_options}
\alias{parse_output_options}
\title{Parse output options from pipeline_options.yaml}
\usage{
parse_output_options(pipeline_options)
}
\arguments{
\item{pipeline_options}{(\code{list}) parsed YAML options}
}
\description{
Parse output options from pipeline_options.yaml
}
\keyword{internal}
