% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{parse_project_name}
\alias{parse_project_name}
\alias{parse_file_extension}
\alias{parse_orient}
\alias{parse_duplicate_policy}
\alias{parse_custom_sample_table}
\alias{assert_supplemental_header_ids}
\alias{parse_supplemental_asv_options}
\alias{parse_added_reference}
\alias{parse_parallel_options}
\alias{parse_forward_primer}
\alias{parse_reverse_primer}
\alias{parse_trim_options}
\alias{parse_unoise_options}
\alias{parse_denoising_options}
\alias{parse_filter_options}
\alias{parse_uncross_options}
\alias{parse_lulu_options}
\alias{parse_amplicon_model_options}
\alias{parse_control_options}
\alias{parse_taxonomy_options}
\alias{parse_taxonomy_ranks}
\alias{parse_protax_options}
\alias{parse_sintax_options}
\alias{parse_bayesant_options}
\alias{parse_epa_options}
\alias{parse_outgroup_options}
\alias{parse_cluster_options}
\alias{parse_cluster_thresholds}
\alias{parse_dist_config}
\alias{parse_guilds_options}
\alias{parse_otu_table_options}
\alias{parse_rarefy_options}
\alias{parse_pipeline_options}
\title{Parse pipeline options from a YAML file}
\usage{
parse_project_name(pipeline_options)

parse_file_extension(pipeline_options)

parse_orient(pipeline_options)

parse_duplicate_policy(pipeline_options)

parse_custom_sample_table(pipeline_options)

assert_supplemental_header_ids(path, set_name)

parse_supplemental_asv_options(pipeline_options)

parse_added_reference(pipeline_options)

parse_parallel_options(pipeline_options)

parse_forward_primer(pipeline_options)

parse_reverse_primer(pipeline_options)

parse_trim_options(pipeline_options)

parse_unoise_options(unoise_options)

parse_denoising_options(pipeline_options)

parse_filter_options(pipeline_options)

parse_uncross_options(pipeline_options)

parse_lulu_options(pipeline_options)

parse_amplicon_model_options(pipeline_options)

parse_control_options(pipeline_options)

parse_taxonomy_options(pipeline_options)

parse_taxonomy_ranks(rank_options)

parse_protax_options(protax_options)

parse_sintax_options(sintax_options)

parse_bayesant_options(bayesant_options)

parse_epa_options(epa_options)

parse_outgroup_options(pipeline_options)

parse_cluster_options(pipeline_options)

parse_cluster_thresholds(thresh_opts)

parse_dist_config(dist_config)

parse_guilds_options(pipeline_options)

parse_otu_table_options(pipeline_options)

parse_rarefy_options(pipeline_options)

parse_pipeline_options()
}
\arguments{
\item{pipeline_options}{(\code{list}) the pipeline options to parse}

\item{set_name}{(\code{character} scalar) name of the supplemental ASV set}

\item{rank_options}{(\code{list} or \code{character} vector) the taxonomic ranks to
use in the pipeline, in order from most inclusive (e.g., kingdom) to least
inclusive (e.g., species). Values may be either named or unnamed. When
named, the name is taken to be the rank, and the value is the "in-group"
taxon at that rank, i.e. the taxon for which results are desired. When
unnamed, the value is taken to be the rank. Example: \code{list(kingdom = "Fungi", "phylum", "class", "order", "family", "genus", "species")}}

\item{protax_options}{(\code{list}) the protax options to parse}

\item{sintax_options}{(\code{list}) the sintax options to parse}

\item{bayesant_options}{(\code{list}) the bayesant options to parse}

\item{epa_options}{(\code{list}) the epa options to parse}
}
\value{
\code{NULL}.  This function is called for its side effect, which is to
configure global options.
}
\description{
These functions parse the pipeline options from a YAML file and set
package options accordingly.
}
\keyword{internal}
