% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protax.R
\name{parse_protax_nameprob}
\alias{parse_protax_nameprob}
\title{Parse the output of ProtaxFungi}
\usage{
parse_protax_nameprob(nameprob, id_is_int = FALSE)
}
\arguments{
\item{nameprob}{(\code{character} string) path to the "nameprob*" files output by
ProtaxFungi}

\item{id_is_int}{(\code{logical}) if \code{TRUE}, sequence labels should be interpreted
as integers}
}
\value{
a \code{data.frame} with columns \code{seq_id} (or \code{seq_idx} if \code{id_is_int} is
\code{TRUE}), \code{rank}, \code{taxonomy}, \code{prob}
}
\description{
Parse the output of ProtaxFungi
}
