% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protax.R
\name{parse_protaxAnimal_output}
\alias{parse_protaxAnimal_output}
\title{Parse the output of ProtaxAnimal}
\usage{
parse_protaxAnimal_output(x)
}
\arguments{
\item{x}{(\code{character}) raw output of ProtaxAnimal, one line per vector element}
}
\value{
a \code{data.frame} with columns \code{seq_id}, \code{rank}, \code{parent_taxonomy},
\code{taxon}, and \code{prob}
}
\description{
Parse the output of ProtaxAnimal
}
