% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_table.R, R/zzz.R
\name{path}
\alias{path}
\alias{meta_path}
\alias{seq_path}
\alias{raw_path}
\alias{rarefy_path}
\alias{trim_path}
\alias{filt_path}
\alias{asv_path}
\alias{aligned_path}
\alias{protax_path}
\alias{output_path}
\alias{log_path}
\alias{paths}
\title{Paths for input, intermediate, and output files}
\usage{
path()

meta_path()

seq_path()

raw_path()

rarefy_path()

trim_path()

filt_path()

asv_path()

aligned_path()

protax_path()

output_path()

log_path()
}
\description{
Paths for input, intermediate, and output files
}
