% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{read_file_extension}
\alias{read_file_extension}
\alias{duplicate_samples}
\alias{custom_sample_table}
\alias{do_custom_sample_table}
\alias{do_supp_asv}
\alias{supp_asv_sets}
\alias{supp_asv_set_names}
\alias{supp_asv_set}
\alias{supp_asv_sequences}
\alias{supp_asv_sample_table}
\alias{supp_asv_taxonomy_mode}
\alias{supp_asv_taxonomy_file}
\alias{do_added_reference}
\alias{added_reference_fasta}
\alias{added_reference_table}
\alias{max_workers}
\alias{min_workers}
\alias{workers_per_seqrun}
\alias{max_batchsize}
\alias{n_seqrun}
\alias{n_orient_seqrun}
\alias{n_workers}
\alias{forward_primer}
\alias{reverse_primer}
\alias{trim_options}
\alias{denoising_method}
\alias{do_dada2}
\alias{do_unoise}
\alias{denoising_pool}
\alias{unoise_alpha}
\alias{unoise_minsize}
\alias{unoise_merge_min_overlap}
\alias{unoise_merge_max_diffs}
\alias{dada2_maxEE}
\alias{do_tag_jump}
\alias{tag_jump_f}
\alias{tag_jump_p}
\alias{do_lulu}
\alias{lulu_dist_type}
\alias{lulu_max_dist}
\alias{lulu_max_gap_length}
\alias{lulu_max_gap_total}
\alias{lulu_min_abundance_ratio}
\alias{lulu_min_cooccurrence_ratio}
\alias{lulu_use_mean_abundance_ratio}
\alias{lulu_dist_config}
\alias{amplicon_model_type}
\alias{amplicon_model_file}
\alias{max_model_start}
\alias{min_model_end}
\alias{min_model_score}
\alias{do_model_filter}
\alias{do_model_align}
\alias{do_model_align_only}
\alias{do_model_filter_only}
\alias{do_model_both}
\alias{do_numt_filter}
\alias{spike_file}
\alias{do_spike}
\alias{pos_control_file}
\alias{do_pos_control}
\alias{do_protax}
\alias{protax_aligned}
\alias{protax_unaligned}
\alias{protax_location}
\alias{do_sintax}
\alias{sintax_ref}
\alias{do_bayesant}
\alias{bayesant_aligned}
\alias{bayesant_ref}
\alias{bayesant_model}
\alias{do_epa}
\alias{epa_ref}
\alias{epa_taxonomy}
\alias{epa_tree}
\alias{epa_params}
\alias{epa_outgroup}
\alias{do_outgroup}
\alias{outgroup_reference}
\alias{outgroup_taxonomy}
\alias{cluster_thresholds}
\alias{do_optimize_thresholds}
\alias{do_optimize_thresholds_self}
\alias{do_optimize_thresholds_reference}
\alias{do_optimize_thresholds_file}
\alias{optimize_thresholds_file}
\alias{cluster_measure}
\alias{cluster_dist_config}
\alias{cluster_force_denovo}
\alias{cluster_min_conf}
\alias{cluster_dist_max}
\alias{cluster_dist_step}
\alias{cluster_min_taxa}
\alias{cluster_min_refseq}
\alias{do_guilds}
\alias{do_wide_otu_table}
\alias{rarefy_number}
\alias{rarefy_numerator}
\alias{rarefy_denominator}
\alias{do_rarefy}
\alias{did_pipeline_options}
\title{Functions to access pipeline-wide options}
\usage{
read_file_extension()

duplicate_samples()

custom_sample_table()

do_custom_sample_table()

do_supp_asv()

supp_asv_sets(enabled_only = TRUE)

supp_asv_set_names(enabled_only = TRUE)

supp_asv_set(set_name)

supp_asv_sequences(set_name)

supp_asv_sample_table(set_name)

supp_asv_taxonomy_mode(set_name)

supp_asv_taxonomy_file(set_name)

do_added_reference()

added_reference_fasta()

added_reference_table()

max_workers()

min_workers()

workers_per_seqrun()

max_batchsize()

n_seqrun()

n_orient_seqrun()

n_workers()

forward_primer()

reverse_primer()

trim_options()

denoising_method()

do_dada2()

do_unoise()

denoising_pool()

unoise_alpha()

unoise_minsize()

unoise_merge_min_overlap()

unoise_merge_max_diffs()

dada2_maxEE()

do_tag_jump()

tag_jump_f()

tag_jump_p()

do_lulu()

lulu_dist_type()

lulu_max_dist()

lulu_max_gap_length()

lulu_max_gap_total()

lulu_min_abundance_ratio()

lulu_min_cooccurrence_ratio()

lulu_use_mean_abundance_ratio()

lulu_dist_config()

amplicon_model_type()

amplicon_model_file()

max_model_start()

min_model_end()

min_model_score()

do_model_filter()

do_model_align()

do_model_align_only()

do_model_filter_only()

do_model_both()

do_numt_filter()

spike_file()

do_spike()

pos_control_file()

do_pos_control()

do_protax()

protax_aligned()

protax_unaligned()

protax_location()

do_sintax()

sintax_ref()

do_bayesant()

bayesant_aligned()

bayesant_ref()

bayesant_model()

do_epa()

epa_ref()

epa_taxonomy()

epa_tree()

epa_params()

epa_outgroup()

do_outgroup()

outgroup_reference()

outgroup_taxonomy()

cluster_thresholds()

do_optimize_thresholds()

do_optimize_thresholds_self()

do_optimize_thresholds_reference()

do_optimize_thresholds_file()

optimize_thresholds_file()

cluster_measure()

cluster_dist_config()

cluster_force_denovo()

cluster_min_conf()

cluster_dist_max()

cluster_dist_step()

cluster_min_taxa()

cluster_min_refseq()

do_guilds()

do_wide_otu_table()

rarefy_number()

rarefy_numerator()

rarefy_denominator()

do_rarefy()

did_pipeline_options()
}
\arguments{
\item{enabled_only}{(\code{logical} flag) if TRUE, only return enabled sets}

\item{set_name}{(\code{character} vector) name(s) of the supplemental ASV set(s)
to access}
}
\value{
(\code{character} string) the forward primer

(\code{character} string) the reverse primer
}
\description{
When used in a targets plan, these should always be
pre-evaluated with \verb{!!} or \verb{!!!} to ensure proper dependency tracking.
}
