% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{prep_table_for_taxsort}
\alias{prep_table_for_taxsort}
\title{Prepare a taxon table for joining with \code{asv_taxsort}}
\usage{
prep_table_for_taxsort(taxon_table, join_by)
}
\arguments{
\item{taxon_table}{(\code{data.frame}) known or closed-reference taxon table.}

\item{join_by}{(\code{character}) \code{"seq_idx_in"} or \code{"seq_id"}.}
}
\value{
(\code{data.frame}) table ready for joining with \code{asv_taxsort}.
}
\description{
Prepare a taxon table for joining with \code{asv_taxsort}
}
\keyword{internal}
