% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{project_name}
\alias{project_name}
\title{Get the currently defined project name}
\usage{
project_name()
}
\value{
(\code{character} string) the project name
}
\description{
Get the currently defined project name
}
