% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{rank2factor}
\alias{rank2factor}
\alias{int2rankfactor}
\title{Convert taxonomic ranks from character vectors and integers to ordered factors}
\usage{
rank2factor(x)

int2rankfactor(x)
}
\arguments{
\item{x}{the taxonomic ranks to convert}
}
\value{
an ordered factor of taxonomic ranks
}
\description{
The least inclusive rank is given the smallest value in the ordering,
so for instance \code{"species" < "kingdom"}.
}
