% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{rank_offset}
\alias{rank_offset}
\alias{set_rank_offset}
\title{Get or set the offset to use when converting between integers and taxonomic
ranks}
\usage{
rank_offset()

set_rank_offset(value)
}
\arguments{
\item{value}{(\code{integer}) the offset to use when converting between integers
and taxonomic ranks}
}
\value{
an integer offset

the previously set offset
}
\description{
Get or set the offset to use when converting between integers and taxonomic
ranks
}
