% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{rarefy_meta}
\alias{rarefy_meta}
\title{Mapping variables for rarefaction}
\usage{
rarefy_meta(dots = TRUE)
}
\value{
a \code{data.frame} giving the variables to map over for rarefaction;
when rarefaction is to be performed (as determined by \code{do_rarefy()}) then
these will always include \code{.rarefy_text} (\verb{character)}), and will also
include either \code{.numerator} (\code{integer}) and \code{.denominator} (\code{integer}) or
\code{.number} (\code{integer}).  Alternatively, if \code{do_rarefy()} returns \code{FALSE}, an
empty data.frame.
}
\description{
Mapping variables for rarefaction
}
