% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{read_hmmer_tblout}
\alias{read_hmmer_tblout}
\alias{read_domtblout}
\alias{read_dna_tblout}
\title{Open a HMMER fixed-width "tblout" file}
\usage{
read_hmmer_tblout(file, col_names, col_types)

read_domtblout(file)

read_dna_tblout(file)
}
\arguments{
\item{file}{(\code{character} string giving file name, or a connection) file to
read}

\item{col_names}{(\code{character} vector) names to apply to columns; passed to
\code{readr::read_fwf()}}

\item{col_types}{(\code{character} string or object returned by \code{readr::cols()})
column typed; passed to \code{readr::read_fwf()}}
}
\value{
a \code{\link[tibble:tibble]{tibble}} giving the contents of the file
}
\description{
This function determines the column widths using the header lines of the
}
\section{Functions}{
\itemize{
\item \code{read_domtblout()}: Read a HMMER domain hits file

\item \code{read_dna_tblout()}: Read a NHMMER hits file

}}
