% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{read_orientation}
\alias{read_orientation}
\title{Get the read orientation setting}
\usage{
read_orientation()
}
\value{
(\code{character} string) the read orientation setting, one of
\code{"fwd"}, \code{"rev"}, \code{"mixed"}, or \code{"custom"}
}
\description{
Get the read orientation setting
}
