% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_table.R
\name{read_sample_table}
\alias{read_sample_table}
\title{Read sample table from file}
\usage{
read_sample_table(sample_table_file = custom_sample_table())
}
\arguments{
\item{sample_table_file}{(\code{character}) the path to the sample table file.}
}
\value{
A \code{tibble} containing the sample table.
}
\description{
This function reads a custom sample table from CSV, TSV, or Excel files.
}
\details{
Required columns in the sample table are \code{sample}, \code{seqrun}, \code{fastq_R1}, and
\code{fastq_R2}. These are all \code{character} vectors, giving the sample name,
sequencing run identifier, and paths to the two read files, respectively.
Optional columns are \code{neg_control} and \code{pos_control}, both of which should be
\code{logical} values (or something which may be interpreted as such) indicating
whether the sample is a negative or positive control, respectively; as well
as \code{orient}, which may by \code{fwd}, \code{rev}, or \code{mixed} to indicate the
orientation of the reads in each sample. Additionally trimming and filtering
parameters may be specified in additional columns, with names matching the
parameter names in \code{pipeline_options.yaml}.
}
\keyword{internal}
