% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infernal.R
\name{read_sfile}
\alias{read_sfile}
\title{Parse the "sfile" written by infernal's cmalign}
\usage{
read_sfile(file)
}
\arguments{
\item{file}{(\code{character}) the path to the file}
}
\value{
a \code{data.frame} with columns \code{idx}, \code{seq_id}, \code{match_len}, \code{cm_from},
\code{cm_to}, \code{trunc}, \code{bit_sc}, \code{avg_pp}, \code{time_band_calc}, \code{time_alignment},
\code{time_total}, and \code{mem_mb}
}
\description{
Parse the "sfile" written by infernal's cmalign
}
