% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{remove_mycobank_number}
\alias{remove_mycobank_number}
\title{Remove mycobank numbers from genus and species names}
\usage{
remove_mycobank_number(taxon)
}
\arguments{
\item{taxon}{(\code{character}) the taxon name to remove the mycobank number from}
}
\value{
the taxon name with the mycobank number removed
}
\description{
Remove mycobank numbers from genus and species names
}
