% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncross.R
\name{remove_tag_jumps}
\alias{remove_tag_jumps}
\title{Remove potential tag-jump from sequence table}
\usage{
remove_tag_jumps(seqtable, f, p, id_col = "seq")
}
\arguments{
\item{seqtable}{(\code{data.frame}) sequence table, as returned by
\code{make_long_sequence_table()} or \code{make_mapped_sequence_table()}}

\item{f}{(\code{numeric}) expected cross-talk rate, e.g. 0.01}

\item{p}{(\code{numeric}) power to rise the exponent (default, 1; use 1/2 or 1/3
to make curves more steep)}

\item{id_col}{(\code{character}) name of the column uniquely identifying the
sequence}
}
\value{
\code{data.frame} with the input columns (including \code{id_col}) plus
\code{total}, \code{uncross}, and \code{is_tag_jump}. Row order matches \code{seqtable}.
}
\description{
Remove potential tag-jump from sequence table
}
