% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_table.R
\name{sample_table}
\alias{sample_table}
\alias{sample_table_hash}
\title{Return the sample table}
\usage{
sample_table(...)

sample_table_hash()
}
\arguments{
\item{...}{arguments to be used for \code{targets} dependency tracking. Ignored
by this function.}
}
\value{
A \code{tibble} with the sample table, as described in
\code{read_sample_table()}
}
\description{
This function reads the custom sample table from disk or infers it from the
raw sequence files, depending on the configuration in
\code{pipeline_options.yaml}. It caches the result to avoid reading the sample
table multiple times in the same R session. It is best to avoid calling this
function on remote workers.
}
