% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_helpers.R
\name{select_sequence}
\alias{select_sequence}
\alias{select_sequence.data.frame}
\alias{select_sequence.character}
\alias{select_sequence.default}
\title{Select elements from a sequence set}
\usage{
select_sequence(seq, which, negate = FALSE, ...)

\method{select_sequence}{data.frame}(seq, which, negate = FALSE, name_col = find_name_col(seq), ...)

\method{select_sequence}{character}(seq, which, negate = FALSE, ...)

\method{select_sequence}{default}(seq, which, negate = FALSE, ...)
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character} string representing a file name, or
any object with named elements) sequences to select from}

\item{which}{(\code{integer}, \code{logical}, or \code{character}) elements to select}

\item{negate}{(\code{logical}) if TRUE, select all elements except those specified}

\item{...}{additional arguments to pass to the selection function}

\item{name_col}{(\code{character} string) name of the column in \code{seq} which
contains the sequence names}
}
\value{
(same type of object as \code{seq}) selected elements
}
\description{
Select elements from a sequence set
}
