% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protax.R
\name{seq_cluster_protax}
\alias{seq_cluster_protax}
\title{Single-linkage clustering of aligned sequences}
\usage{
seq_cluster_protax(aln_seq, aln_index, which, thresh, aln_len)
}
\arguments{
\item{aln_seq}{(\code{character}) file name of gzipped FASTA file containing
sequences to cluster}

\item{aln_index}{(\code{character} or \code{fastqindexr_index}) file path or
\code{\link[fastqindexr:create_index]{fastqindexr_index}} object for \code{aln_seq}}

\item{which}{(\code{integer}) integer indices of sequences to cluster}

\item{thresh}{(\code{numeric}) clustering similarity threshold as a percentage,
i.e., 0.0 to 100.0}

\item{aln_len}{(\code{integer}) width of alignment}
}
\value{
a \code{data.frame} with columns \code{seq_id}, \code{cluster}, and \code{dist}
}
\description{
Single-linkage clustering of aligned sequences
}
