% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_helpers.R
\name{sequence_size}
\alias{sequence_size}
\alias{sequence_size.XStringSet}
\alias{sequence_size.character}
\alias{sequence_size.default}
\title{Get the number of sequences in a sequence set}
\usage{
sequence_size(seq, ...)

\method{sequence_size}{XStringSet}(seq, ...)

\method{sequence_size}{character}(seq, ...)

\method{sequence_size}{default}(seq, ...)
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character} string representing a file name,
\code{\link[Biostrings:XStringSet-class]{XStringSet}}, or any object with multiple elements)
sequences to count}

\item{...}{additional arguments to pass to methods}
}
\value{
(\code{integer}) number of sequences
}
\description{
Get the number of sequences in a sequence set
}
