% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{sintax_format}
\alias{sintax_format}
\title{Format a classification for use as a Sintax reference DB}
\usage{
sintax_format(s)
}
\arguments{
\item{s}{(\code{character} vector) the classification to format; should be
comma-delimited from most inclusive rank to least inclusive}
}
\description{
Format a classification for use as a Sintax reference DB
}
