% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_table.R
\name{sort_seq_table}
\alias{sort_seq_table}
\alias{sort_seq_table.matrix}
\alias{sort_seq_table.data.frame}
\title{Sort a sequence table}
\usage{
sort_seq_table(seqtable, ...)

\method{sort_seq_table}{matrix}(seqtable, ...)

\method{sort_seq_table}{data.frame}(seqtable, seqs = NULL, abund_col = "nread", ...)
}
\arguments{
\item{seqtable}{(\code{matrix} or \code{data.frame}) a sequence table as returned by
\code{dada2::makeSequenceTable()}, \code{make_long_sequence_table()}, or
\code{make_mapped_sequence_table()}}

\item{...}{additional arguments passed to methods, currently ignored}

\item{seqs}{(\code{character}, \code{\link[Biostrings:XStringSet-class]{DNAStringSet}},
file name, or \code{data.frame}) sequences, for mapped sequence tables (currently
unused)}

\item{abund_col}{(\code{character}) column name containing abundance information}
}
\value{
for the \code{matrix} method, a sorted sequence table; for the \code{data.frame}
method, appropriate identifiers for the sequences, sorted.  These will be
either the sequences themselves, \code{seq_id}, or \code{seq_idx}, depending on the
input.
}
\description{
Sort a sequence table
}
