% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncross.R
\name{summarize_uncross}
\alias{summarize_uncross}
\title{Summarize uncrossing information}
\usage{
summarize_uncross(uncross)
}
\arguments{
\item{uncross}{(\code{data.frame}) uncrossing information, as returned by
\code{remove_tag_jumps()}.}
}
\value{
\code{data.frame} with columns \code{sample}, \code{Total_reads},
\code{Number_of_TagJump_Events}, \code{TagJump_reads}, and \code{ReadPercent_removed}.
}
\description{
Summarize uncrossing information
}
