% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_map_bind_rows}
\alias{tar_map_bind_rows}
\alias{tar_map_c}
\alias{tar_map_list}
\title{Generate quosures which combines static branching targets}
\usage{
tar_map_bind_rows(plan, target_name = NULL)

tar_map_c(plan, target_name = NULL)

tar_map_list(plan, target_name = NULL)
}
\arguments{
\item{plan}{(a named, nested \code{list} of \code{targets::tar_target()} objects, as
produced by \code{targets::tar_map()}) the targets plan to operate on}

\item{target_name}{(\code{character} string) the pre-mapping name of the target
to combine}
}
\value{
a \code{quosure} which combines the targets
}
\description{
Generate quosures which combines static branching targets

Generate a quosure which combines static branching targets with \code{vctrs::vec_c()}

Generate a quosure which combines static branching targets with \code{list()}
}
