% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_map_symbols}
\alias{tar_map_symbols}
\title{Get variants of a target name which has been run through "tar_map"}
\usage{
tar_map_symbols(plan, target_name = NULL)
}
\arguments{
\item{plan}{(a named nested \code{list} of \code{targets::tar_target()} objects, as
produced by \code{targets::tar_map()}) the targets plan to extract target names
from}

\item{target_name}{(\code{character} string) the pre-mapping name of the target
to extract variants of}
}
\value{
a \code{list} of \code{symbol} giving the names of the target variants
}
\description{
Get variants of a target name which has been run through "tar_map"
}
