% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_merge}
\alias{tar_merge}
\title{Merge named lists within two targets (sub)plans}
\usage{
tar_merge(plan1, plan2)
}
\arguments{
\item{plan1}{(\code{targets::tar_target()} object or (possibly nested) list of such
objects) the first plan to merge}

\item{plan2}{(\code{targets::tar_target()} object or (possibly nested) list of such
objects) the second plan to merge}
}
\value{
a nested \code{list} of \code{targets::tar_target()} objects
}
\description{
This is most meaningful when the plans were both produced by \code{targets::tar_map()}
and the targets in the plans have the same names.
}
