% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{tar_substitute}
\alias{tar_substitute}
\title{Substitute values into a target object or plan}
\usage{
tar_substitute(target, values)
}
\arguments{
\item{target}{(\link[targets:tar_target]{target} object) the target(s) to substitute}

\item{values}{(\code{list}) a named list of values to substitute into the target}
}
\value{
(\link[targets:tar_target]{target} object) the target with substituted values
}
\description{
Substitute values into a target object or plan
}
