% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tax_opts}
\alias{tax_opts}
\alias{root_rank}
\alias{root_rank_var}
\alias{second_rank}
\alias{second_rank_var}
\alias{ingroup_rank}
\alias{ingroup_rank_var}
\alias{ingroup_taxon}
\alias{tip_rank}
\alias{tip_rank_var}
\title{Shortcuts for accessing taxonomy options}
\usage{
root_rank()

root_rank_var()

second_rank()

second_rank_var()

ingroup_rank()

ingroup_rank_var()

ingroup_taxon()

tip_rank(ranks = tax_ranks())

tip_rank_var(ranks = tax_ranks())
}
\description{
Shortcuts for accessing taxonomy options
}
\section{Functions}{
\itemize{
\item \code{root_rank()}: Get the root taxonomic rank as a string

\item \code{root_rank_var()}: Get the root taxonomic rank as a symbol

\item \code{second_rank()}: Get the second taxonomic rank as a string

\item \code{second_rank_var()}: Get the second taxonomic rank as a symbol

\item \code{ingroup_rank()}: Get the ingroup taxonomic rank as a string

\item \code{ingroup_rank_var()}: Get the ingroup taxonomic rank as a symbol

\item \code{ingroup_taxon()}: Get the ingroup taxon as a string

\item \code{tip_rank()}: Get the tip taxonomic rank as a string

\item \code{tip_rank_var()}: Get the tip taxonomic rank as a symbol

}}
