% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tax_table_wide_to_long}
\alias{tax_table_wide_to_long}
\title{Convert wide taxonomy/probability tables to long format}
\usage{
tax_table_wide_to_long(taxonomy, probability, ranks = tax_ranks())
}
\arguments{
\item{taxonomy}{(\code{data.frame}) wide table with one taxon column per rank and
a \code{seq_id} column.}

\item{probability}{(\code{data.frame}) wide table with one probability column per
rank and a \code{seq_id} column.}

\item{ranks}{(\code{character} vector) taxonomic ranks to pivot.}
}
\value{
\code{tibble} with columns \code{seq_id}, \code{rank}, \code{parent_taxonomy}, \code{taxon},
and \code{prob}.
}
\description{
Convert wide taxonomy/probability tables to long format
}
