% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{threshold_as_dist}
\alias{threshold_as_dist}
\title{Convert thresholds which may be distances or similarities to distances}
\usage{
threshold_as_dist(thresholds)
}
\arguments{
\item{thresholds}{(\code{numeric}) thresholds to convert. These may be distances
(0 is identity) with values between 0.0 and 0.5, or between 0.0 and 50.0,
in which case they are interpreted as percentages;
or similarities (1 or 100 is identity) with values between 0.5 and 1.0, or
between 50.0 and 100.0. Note that if \emph{all} percentage distances are < 0.5,
they will be misinterpreted as fractional distances.}
}
\value{
(\code{numeric}) thresholds as distances, in the range from 0.0 to 1.0
}
\description{
Convert thresholds which may be distances or similarities to distances
}
