% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R, R/pipeline_options.R
\name{trim_primer}
\alias{trim_primer}
\alias{trim_primer_R1}
\alias{trim_primer_R2}
\alias{trim_primer_merged}
\title{Trim primers from a set of sequences}
\usage{
trim_primer(seqs, primer, ...)

trim_primer_R1()

trim_primer_R2()

trim_primer_merged()
}
\arguments{
\item{seqs}{(\code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, \code{character}
vector, or \code{data.frame}) sequences to trim}

\item{primer}{(\code{character}) primer sequence}

\item{...}{additional arguments to \code{cutadapt_filter_trim()}}
}
\value{
a \code{tibble} with columns \code{seq_id} and \code{seq}

(\code{character}) string with the primer setting
}
\description{
Trim primers from a set of sequences

Get the combined primer string for trimming reads
}
