% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{truncate_taxonomy}
\alias{truncate_taxonomy}
\title{Truncate classification(s) at a given rank}
\usage{
truncate_taxonomy(s, rank)
}
\arguments{
\item{s}{(\code{character} vector`) the classification(s) to truncate}

\item{rank}{(\code{integer}) the rank to truncate at}
}
\value{
a \code{character} vector of truncated classifications
}
\description{
Truncate classification(s) at a given rank
}
