% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{update.dada2_filter_options}
\alias{update.dada2_filter_options}
\title{Update method for dada2 filter options}
\usage{
\method{update}{dada2_filter_options}(object, new_options, ...)
}
\arguments{
\item{object}{(\code{\link{dada2_filter_options}}) existing
dada2 filter options object to modify}

\item{new_options}{(named \code{list}, single-row \code{data.frame}, named
\code{character} vector, or named \code{numeric} vector) new values for the options.
If a \code{data.frame}, then the values in each column should be all the same.}

\item{...}{Additional arguments (ignored)}
}
\description{
Update method for dada2 filter options
}
