% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{update.merged_filter_options}
\alias{update.merged_filter_options}
\title{Update method for merged filter options}
\usage{
\method{update}{merged_filter_options}(object, new_options, ...)
}
\arguments{
\item{object}{(\code{\link{merged_filter_options}}) existing
merged filter options object to modify}

\item{new_options}{(named \code{list}, single-row \code{data.frame}, named
\code{character} vector, or named \code{numeric} vector) new values for the options.
If a \code{data.frame}, then the values in each column should be all the same.}

\item{...}{Additional arguments (ignored)}
}
\description{
Update method for merged filter options
}
