% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\name{update.cutadapt_paired_options}
\alias{update.cutadapt_paired_options}
\alias{update.cutadapt_options}
\title{Replace existing cutadapt options with new values}
\usage{
\method{update}{cutadapt_paired_options}(options, new_options)

\method{update}{cutadapt_options}(options, new_options)
}
\arguments{
\item{options}{(\href{`cutadapt_paired_options()`}{\code{cutadapt_paired_options}} or
\href{`cutadapt_options()`}{\code{cutadapt_options}}) existing cutadapt options to
update}

\item{new_options}{(\code{data.frame}, named \code{list}, or named \code{character} vector)
new values for the options. If a \code{data.frame}, then the values in the
\code{data.frame} should be all the same.}
}
\value{
updated options
}
\description{
Replace existing cutadapt options with new values
}
