% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{validate_asv_taxsort}
\alias{validate_asv_taxsort}
\title{Validate \code{asv_taxsort} column layout}
\usage{
validate_asv_taxsort(asv_taxsort)
}
\arguments{
\item{asv_taxsort}{(\code{data.frame}) ASV taxonomic sort table.}
}
\value{
(\code{character}) join key: \code{"seq_idx_in"} or \code{"seq_id"}.
}
\description{
Validate \code{asv_taxsort} column layout
}
\keyword{internal}
