% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{vsearch_cluster_unoise2}
\alias{vsearch_cluster_unoise2}
\title{Find denoised sequences using UNOISE3 (\code{vsearch --cluster_unoise})}
\usage{
vsearch_cluster_unoise2(
  seq,
  min_size = 8,
  alpha = 2,
  threads = 1,
  shards = min(local_cpus()\%/\%threads, length(seq)),
  vsearch = find_vsearch()
)
}
\arguments{
\item{seq}{(\code{character} vector) path(s) of one or more FASTA or FASTQ files,
possibly gzipped}

\item{min_size}{(\code{integer} scalar) minimum abundance of a sequence to be
considered a centroid (\code{--minsize})}

\item{alpha}{(\code{numeric} scalar) alpha parameter for UNOISE (\code{--unoise_alpha})}

\item{threads}{(\code{integer} scalar) number of threads to use per vsearch
process}

\item{shards}{(\code{integer} scalar) number of shards (parallel vsearch calls) to
use}

\item{vsearch}{(\code{character} scalar) path to vsearch executable}
}
\value{
a named \code{list} of objects of class \code{uc_cluster}, where names match
the values of \code{seq}. Each \code{uc_cluster} has two members, each a
\code{data.frame}:
\itemize{
\item \code{clusters} with columns:
\itemize{
\item \code{clust_idx} - cluster index; integer from 0 to number of clusters - 1
\item \code{size} - the number of reads in the cluster
\item \code{seq} - the representative sequence (i.e. centroid) of the cluster
}
\item \code{map} with columns:
\itemize{
\item \code{clust_idx} - cluster index; integer from 0 to number of clusters - 1
\item \code{seq_id} - the sequence ID of a sequence which is a member of the
cluster. Although input sequences are typically dereplicated and
include a \code{";size="} annotation, it is stripped off so that the ID
should match a read in the original file.
}
}

Empty inputs return an empty \code{uc_cluster} without invoking vsearch.
}
\description{
This command actually executes both \code{--fastx_uniques} (or equivalent
dereplication) and \code{--cluster_unoise}, since the former is required for the
latter. The function name retains "unoise2" for historical reasons.
}
