% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{vsearch_fastq_merge_pairs}
\alias{vsearch_fastq_merge_pairs}
\title{Assemble Illumina read pairs using USEARCH or VSEARCH}
\usage{
vsearch_fastq_merge_pairs(
  seq_R1,
  seq_R2,
  seq_out,
  min_overlap = 5,
  max_mismatch = 10,
  threads = 1,
  shards = min(local_cpus()\%/\%threads, length(seq_out)),
  filter_options = merged_filter_options(),
  compress = all(endsWith(seq_out, ".gz")),
  fastq = all(grepl("[.]fa?s?t?q", basename(seq_out))),
  vsearch = find_vsearch()
)
}
\arguments{
\item{seq_R1}{(\code{character} vector) path(s) of one or more FASTQ files,
possibly gzipped, to be assembled in the forward orientation.}

\item{seq_R2}{(\code{character} vector) path(s) of one or more FASTQ files,
possibly gzipped, to be assembled in the reverse complement orientation.
Must be the same number of files as \code{seq_R1}, and each file must have the
same number of reads.}

\item{seq_out}{(\code{character} vector) file names}

\item{min_overlap}{(\code{integer} scalar) minimum length of the overlapping
region in order to assemble a read pair.}

\item{max_mismatch}{(\code{numeric} scalar) if strictly < 1, the fraction of bases
in the overlapping region which are allowed to be mismatches (
\code{--fastq_maxdiffpct} argument, but should be expressed as a fraction rather
than percent). If >= 1, the number of bases in the overlapping region which
are allowed to be mismatches (\code{--fastq_maxdiffs} argument). In the latter
case it should be integer-valued. Default: 10}

\item{threads}{(\code{integer} scalar) number of threads to use for \emph{each}
usearch/vsearch process. Individual processes eventually become I/O bound
and more threads may provide little improvement, or even slow the total
processing time. The point of diminishing returns is probably dependent on
the system and the input size.}

\item{shards}{(\code{integer} scalar) number of parallel usearch/vsearch processes
to run at once.  The total number of CPU threads used is something like
\code{threads * shards} (although the R process itself will also consume some).}

\item{filter_options}{(result object from \code{\link[=merged_filter_options]{merged_filter_options()}}) additional
options to filter the results; passed to USEARCH or VSEARCH.}

\item{compress}{(\code{logical} flag) if \code{TRUE}, then vsearch/usearch output is
piped through \code{gzip} to compress it. The default autodetects based on
whether the filenames in \code{seq_out} end with \code{".gz"}.}

\item{fastq}{(\code{logical} flag) if \code{TRUE}, then output is in FASTQ format;
otherwise it is FASTA. By default this is autodetected from the filenames
in \code{seq_out}.}

\item{vsearch}{(\code{character} string) path to USEARCH or VSEARCH executable.}
}
\value{
\code{character} vector giving the output file names (as given in
\code{seq_out})
}
\description{
Assemble Illumina read pairs using USEARCH or VSEARCH
}
