% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_and_return_file.R
\name{write_and_return_file}
\alias{write_and_return_file}
\alias{write_and_return_file.XStringSet}
\alias{write_and_return_file.data.frame}
\alias{write_and_return_file.matrix}
\alias{write_and_return_file.list}
\alias{write_and_return_file.character}
\alias{write_and_return_file.ggplot}
\alias{write_and_return_file.default}
\title{Write an object to a file and return the file path}
\usage{
write_and_return_file(x, file, ...)

\method{write_and_return_file}{XStringSet}(x, file, width = 20001L, ...)

\method{write_and_return_file}{data.frame}(x, file, type = NULL, ...)

\method{write_and_return_file}{matrix}(x, file, type = NULL, ...)

\method{write_and_return_file}{list}(x, file, type = NULL, envir = parent.frame(), ...)

\method{write_and_return_file}{character}(x, file, ...)

\method{write_and_return_file}{ggplot}(x, file, ...)

\method{write_and_return_file}{default}(x, file, type = NULL, ...)
}
\arguments{
\item{x}{(any object) object to write}

\item{file}{(\code{character} string) file path to write to}

\item{...}{additional arguments passed to methods; several methods accept
a \code{type} argument (see method documentation)}

\item{width}{(integer) the maximum width of each line in the output file}

\item{type}{(\code{character}) tabular format: \code{"rds"}, \code{"tsv"}, \code{"csv"},
\code{"xlsx"}, \code{"fst"}, \code{"feather"}, \code{"parquet"}, \code{"qs2"}, or \code{"qdata"}
(\code{"qd"} is an alias). If \code{NULL}, inferred from \code{file} extension (not
case-sensitive).}

\item{envir}{(\code{environment}) environment for resolving symbol names when
writing RData from an unnamed list of names or symbols}
}
\value{
the file path
}
\description{
Write an object to a file and return the file path
}
