% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_helpers.R
\name{write_sequence}
\alias{write_sequence}
\alias{write_sequence.data.frame}
\alias{write_sequence.character}
\alias{write_sequence.XStringSet}
\title{Write sequences as a fasta file}
\usage{
write_sequence(seq, fname, ...)

\method{write_sequence}{data.frame}(
  seq,
  fname,
  seq_col = find_seq_col(seq),
  name_col = find_name_col(seq),
  ...
)

\method{write_sequence}{character}(seq, fname, ...)

\method{write_sequence}{XStringSet}(seq, fname, ...)
}
\arguments{
\item{seq}{(\code{data.frame}, \code{character}, or \code{XStringSet}) sequences to write}

\item{fname}{(\code{character} string) file path to write to}

\item{...}{additional arguments to pass to the writing function}

\item{seq_col}{(\code{character} string) name of the column in \code{seq} which
contains the sequences}

\item{name_col}{(\code{character} string) name of the column in \code{seq} which
contains the sequence names}
}
\description{
Write sequences as a fasta file
}
