% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krona.R
\name{xml_format}
\alias{xml_format}
\title{Internal function: convert a list of data to the XML format to be sent to KronaTools}
\usage{
xml_format(data_format)
}
\arguments{
\item{data_format}{(\code{list}) a named list whose elements are values to be
formatted as data for a KronaTools XML file}
}
\value{
(\code{character}) the XML-formatted data
}
\description{
Converts \code{list(type1 = c(val1a, val1b), type2 = c(val2a, val2b))} to
\verb{<type1><val>\{val1a\}</val><val>\{val1b\}</val></type1><type2><val>\{val2a\}</val><val>\{val2b\}</val></type2>}
}
