% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncross.R
\name{add_uncross_to_seq_map}
\alias{add_uncross_to_seq_map}
\title{Add uncrossing information to a sequence map}
\usage{
add_uncross_to_seq_map(seqmap, seqtable_raw, uncross)
}
\arguments{
\item{seqmap}{(\code{data.frame}) sequence map, as returned by \code{seq_map()}}

\item{seqtable_raw}{(\code{data.frame}) raw sequence table, as returned by
\code{make_mapped_sequence_table()}.}

\item{uncross}{(\code{data.frame}) uncrossing information, as returned by
\code{remove_tag_jumps()}.}
}
\value{
\code{data.frame} with the same columns as \code{seqmap}, but with the \code{flags}
column updated to include the \code{is_tag_jump} information from \code{uncross}.
}
\description{
Add uncrossing information to a sequence map
}
