% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesant.R
\name{bayesant}
\alias{bayesant}
\title{Identify sequences using BayesANT}
\usage{
bayesant(
  query,
  model,
  ncpu = local_cpus(),
  id_is_int = FALSE,
  n_top_taxa = 20,
  min_prob = 0.01
)
}
\description{
Identify sequences using BayesANT
}
