% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_taxon_thresholds.R
\name{calc_taxon_thresholds}
\alias{calc_taxon_thresholds}
\title{Calculate clustering thresholds for each taxon, falling back to its ancestor
taxa as necessary}
\usage{
calc_taxon_thresholds(
  rank,
  taxon_table,
  optima,
  conf_level = NULL,
  metric = NULL,
  default = ingroup_taxon()
)
}
\arguments{
\item{rank}{(\code{character} string) the rank within which clustering will be
performed}

\item{taxon_table}{(\code{data.frame}) "wide" taxonomy table; column "seq_id"
gives the sequence ID, and columns {\code{root_rank()}} to {\code{tip_rank()}} (e.g.,
"kingdom" to "species") give the taxonomy at each rank.}

\item{optima}{(\code{data.frame}) optimum clustering thresholds within
various taxa;
column "rank" gives the rank which is approximated by
clustering;
"superrank" gives the rank of the taxon within which the clustering threshold
was optimized;
"supertaxon" gives that taxon name;
"threshold" gives the optimum clustering threshold;
"value" gives the value of the optimization target at the optimum threshold;
optionally, "conf_level" gives a string description of the confidence level;
optionally, "metric" gives a string description of the optimization target.
Typically, "conf_level" and "measure" are only needed when optimization has
been performed for multiple confidence levels or measures, and so the
arguments by the same name should also be included.}

\item{conf_level}{(\code{character} string or NULL) if given, then only rows of
\code{optima} with this value in the "conf_level" column will be used}

\item{metric}{(\code{character} string or NULL) if given, then only rows of
\code{optima} with this value in the "metric" column will be used}

\item{default}{(\verb{character string}) default taxon to define threshold to use
when taxonomy is unknown. default: \code{ingroup_taxon()}}
}
\value{
(named \code{numeric}, where names are taxa and values are clustering
thresholds)
unknown sequences, grouped by taxonomy at the parent rank. Sequences where
the parent rank is also unknown are in the item named \code{"_NA_"}.
}
\description{
Calculate clustering thresholds for each taxon, falling back to its ancestor
taxa as necessary
}
