% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrappers.R
\name{choose_dada_error_function}
\alias{choose_dada_error_function}
\title{Choose error function for dada2}
\usage{
choose_dada_error_function(fls, ...)
}
\arguments{
\item{fls}{File names}
}
\value{
A function that estimates error rates based on binned quality scores
if the input files are binned, otherwise the default dada2 error function.
}
\description{
Choose error function for dada2
}
