% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infernal.R
\name{consensus_columns}
\alias{consensus_columns}
\title{Extract the consensus columns from a multiple sequence alignment}
\usage{
consensus_columns(aln)
}
\arguments{
\item{aln}{(\code{list}) a list with elements \code{alignment} (a \code{MultipleAlignment}
object) and \code{GC} (a \code{BString} object with the consensus sequence)}
}
\value{
a \code{DNAStringSet} with the consensus columns
}
\description{
Extract the consensus columns from a multiple sequence alignment
}
