% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numt.R
\name{detect_numts}
\alias{detect_numts}
\title{Detect Nuclear Mitochondrial Paralogs (NUMTs)}
\usage{
detect_numts(a2m, id_is_int = FALSE, files = NULL, seq_idx = NULL, ...)
}
\arguments{
\item{a2m}{(\code{character}) name of an A2M file containing a multiple sequence
alignment as generated by \code{hmmalign()}.  In particular, insertions relative
to the reference alignment should be present as masked (lower-case)
characters.}

\item{id_is_int}{(\code{logical}) if \code{TRUE}, return the sequence index as an
integer}
}
\value{
\code{data.frame} with columns \code{seq_id} (or \code{seq_idx} if \code{id_is_int} is
\code{TRUE}), \code{numt_indicator}, \code{pos}, and \code{len}, where \code{numt_indicator} is one of
"frameshift" or "stop_codon", \code{pos} is the 1-based position of the indicator,
and \code{len} is the length of the indicator.
}
\description{
Detect Nuclear Mitochondrial Paralogs (NUMTs)
}
