% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{do_closed_ref_cluster}
\alias{do_closed_ref_cluster}
\title{Run closed-reference clustering}
\usage{
do_closed_ref_cluster(
  preclosed_taxon_table,
  rank,
  parent_rank,
  seq_file,
  seq_file_index,
  thresholds,
  dist_config,
  parallel_config
)
}
\arguments{
\item{preclosed_taxon_table}{(\code{data.frame}) a table of sequences for which
closed-reference clustering is required. Must have columns \code{seq_idx}
(\code{integer}) giving the indices of sequences in \code{seq_file}, as well as
columns with names matching \code{rank} and \code{parent_rank}.}

\item{rank}{(\code{character}) the taxonomic rank to calculate clusters for}

\item{parent_rank}{(\code{character}) the taxonomic rank to use as a parent}

\item{seq_file}{(\code{character}) the path to the sequence file}

\item{seq_file_index}{(\code{character}) the path to the sequence file index}

\item{thresholds}{(\code{numeric}) a named vector of thresholds to use for
closed-reference clustering. These are interpreted as distances on the
interval \link{0, 100}, where 0 is a perfect match and 100 is a perfect
mismatch.}

\item{dist_config}{(\code{list}) a list of configuration options for the
distance calculation, as returned by \code{\link[optimotu:dist_config]{optimotu::dist_config()}}.}

\item{parallel_config}{(\code{list}) a list of configuration options for the
parallelization, as returned by \code{\link[optimotu:parallel_config]{optimotu::parallel_config()}}.}
}
\value{
(\code{data.frame}) a two-column table of mapping between query
sequences (those where the \code{rank} column in \code{preclosed_taxon_table} is \code{NA})
and reference sequences (those where the \code{rank} column is not \code{NA}).
}
\description{
This function calls \code{\link[optimotu:closed_ref_cluster]{optimotu::closed_ref_cluster()}} for one or more
parent taxa.
}
