% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_and_return_file.R
\name{ensure_directory}
\alias{ensure_directory}
\title{Create the parent directory of a file if it does not exist}
\usage{
ensure_directory(file)
}
\arguments{
\item{file}{(\code{character} string) file path to create the parent directory for}
}
\description{
Create the parent directory of a file if it does not exist
}
