% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_sample}
\alias{fastq_sample}
\title{Sample a given number of reads from a FASTQ file}
\usage{
fastq_sample(infile, outfile, n, sample, rename = FALSE)
}
\arguments{
\item{infile}{(\code{character} string) the path to the input FASTQ file. May be gzipped.}

\item{outfile}{(\code{character} string) the path to the output FASTQ file. If it ends in ".gz", the output will be gzipped.}

\item{n}{(\code{integer}) the number of reads to sample.}

\item{sample}{(\code{integer} vector) a shuffle of the read indices.}

\item{rename}{(\code{logical}) if \code{TRUE}, the read names will be replaced with a hexadecimal sequential
number.}
}
\value{
the output file name
}
\description{
This function samples the given number of reads from a FASTQ file, based on
a shuffle of the read indices. This guarantees reproducible sampling when the
same shuffle is reused.
}
