% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_sample_fraction}
\alias{fastq_sample_fraction}
\title{Subsample from a FASTQ file}
\usage{
fastq_sample_fraction(file, numerator, denominator, output, rename = FALSE)
}
\arguments{
\item{file}{(\code{character} string) the path to the FASTQ file.}

\item{numerator}{(\code{integer} scalar) numerator of the fraction of reads to
sample.}

\item{denominator}{(\code{integer} scalar) denominator of the fraction of reads
to sample.}

\item{output}{(\code{character} string) the path to the output file. If it ends
in ".gz", the output will be gzipped.}

\item{rename}{(\code{logical} scalar) whether to rename the reads in the output;
if \code{TRUE}, the read names will be replaced with a hexadecimal sequential
number}
}
\value{
the output file name
}
\description{
This function subsamples an (optionally gzipped) FASTQ file by randomly
selecting a specified number of reads. The file is streamed and only
\code{denominator} reads are stored in memory at any time, which allows it to
handle very large files without running out of memory. Sampling is
performed with reservoir sampling.
}
\details{
Additionally, it guarantees that, if the same file is resampled with the
same seed, but a larger number of reads, the result from the first sampling
will be a subset of the new, larger sampling.

When the number of sequences in the input file is not divisible by the
denominator, the remainder sequences will be sampled as if there were enough
additional sequences to fill the denominator, so that the actual number of
sampled sequences may differ from (file size) * \code{numerator} / \code{denominator}
by up to max(\code{numerator}, \code{denominator}/2).
}
