% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastq_sample_number}
\alias{fastq_sample_number}
\title{Subsample a FASTQ file to a fixed number of reads}
\usage{
fastq_sample_number(file, number, output, rename = FALSE)
}
\arguments{
\item{file}{(\code{character} string) the path to the input FASTQ file. May be gzipped.}

\item{number}{(\code{integer} scalar) the number of reads to sample.}

\item{output}{(\code{character} string) the path to the output file. If it ends in ".gz", the output will be gzipped.}

\item{rename}{(\code{logical} scalar) whether to rename the reads in the output;
if \code{TRUE}, the read names will be replaced with a hexadecimal sequential
number}
}
\value{
the output file name
}
\description{
This function subsamples a FASTQ file to a fixed number of reads.
}
