% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_index.R
\name{fastx_gz_index}
\alias{fastx_gz_index}
\alias{fastx_gz_extract}
\alias{fastx_gz_random_access_extract}
\title{Indexing and random access extraction of gzipped FASTA and FASTQ files}
\usage{
fastx_gz_index(file)

fastx_gz_extract(
  infile,
  index,
  i,
  outfile,
  renumber = FALSE,
  append = FALSE,
  hash = NULL
)

fastx_gz_random_access_extract(
  infile,
  index,
  i,
  outfile = NULL,
  renumber = FALSE,
  append = FALSE,
  hash = NULL,
  max_gap = 100L,
  ncpu = local_cpus()
)
}
\arguments{
\item{file}{(\code{character} filename) file to create an index for

These functions are wrappers around the
\href{https://dkfz-odcf.github.io/FastqIndEx/}{\code{FastqIndEx}} utility.}

\item{infile}{(\code{character} filename) gzipped fasta or fastq file}

\item{index}{(\code{character} filename) index file for \code{infile}}

\item{i}{(\code{integer} vector) indices to extract}

\item{outfile}{(\code{character} filename) file to write the extracted sequences to}

\item{renumber}{(\code{logical} flag) if \code{TRUE}, replace the sequence names with
integers, starting at 0.}

\item{append}{(\code{logical} flag) if \code{TRUE}, append to \code{outfile} if it already
exists, rather than overwriting.}

\item{hash}{(\code{character} scalar) md5 hash of the infile; ignored but included
as a parameter for dependency tracking}

\item{max_gap}{(\code{integer} scalar) maximum gap between consecutive indices for
a single extraction process}

\item{ncpu}{(\code{integer} scalar) number of processes to use}
}
\value{
file name of the created index

filename of the output file

filename of the output file
}
\description{
Indexing and random access extraction of gzipped FASTA and FASTQ files
}
\section{Functions}{
\itemize{
\item \code{fastx_gz_index()}: Generate a gzip index file

\item \code{fastx_gz_extract()}: Extract sequences from a gzipped FASTA or FASTQ file

\item \code{fastx_gz_random_access_extract()}: Extract sequences from a gzipped FASTA or FASTQ file

}}
