% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.R
\name{find_executable}
\alias{find_executable}
\title{Find an executable}
\usage{
find_executable(executable)
}
\arguments{
\item{executable}{(\code{character} string) the name of the executable to find.}
}
\value{
(\code{character} string) the full path to the executable.
}
\description{
This function tries to find an executable in the system. It first checks the
environment variables, then the system path, and finally the current working
directory.
}
