% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{full_preclosed_taxon_table}
\alias{full_preclosed_taxon_table}
\title{Calculate a full preclosed taxon table}
\usage{
full_preclosed_taxon_table(
  known_taxon_table,
  asv_taxsort,
  rank,
  parent_rank,
  tax_ranks
)
}
\arguments{
\item{known_taxon_table}{(\code{data.frame}) a table of known taxon assignments}

\item{asv_taxsort}{(\code{data.frame}) a table of ASV taxonomic assignments}

\item{rank}{(\code{character}) the taxonomic rank to calculate}

\item{parent_rank}{(\code{character}) the taxonomic rank to use as a parent}

\item{tax_ranks}{(\code{character}) the taxonomic ranks to use}
}
\value{
(\code{data.frame}) a table of sequences for which closed-reference
clustering is required. This includes both sequences which are not assigned
to the target rank which will act as queries, and identified sequences
which belong to the same parent rank as one or more queries. The results
includes a \code{tar_group} column for dynamic branching over each parent
taxon. If there are no queries, a single row is returned to branch over.
}
\description{
This calculates the full table of inputs to be passed to
\code{\link[=do_closed_ref_cluster]{do_closed_ref_cluster()}}, with a single group for each parent taxon.
It is also used internally by \code{\link[=large_preclosed_taxon_table]{large_preclosed_taxon_table()}} and
\code{\link[=small_preclosed_taxon_table]{small_preclosed_taxon_table()}}.
}
