% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{full_predenovo_taxon_table}
\alias{full_predenovo_taxon_table}
\title{Calculate a full pre-denovo taxon table}
\usage{
full_predenovo_taxon_table(
  closedref_taxon_table,
  asv_taxsort,
  rank,
  parent_rank,
  tax_ranks
)
}
\description{
This calculates the full table of inputs to be passed to
\code{\link[=do_denovo_cluster]{do_denovo_cluster()}}, with a single group for each parent taxon.
It is also used internally by \code{\link[=large_predenovo_taxon_table]{large_predenovo_taxon_table()}} and
\code{\link[=small_predenovo_taxon_table]{small_predenovo_taxon_table()}}.
}
