% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\name{gappa_assign}
\alias{gappa_assign}
\title{Run Gappa}
\usage{
gappa_assign(
  jplace,
  taxonomy,
  outgroup,
  ranks = unknown_ranks(),
  ncpu = NULL,
  allow_file_overwriting = TRUE,
  verbose = FALSE,
  id_is_int = FALSE
)
}
\arguments{
\item{jplace}{(\code{character}) path to the jplace file or a list containing
jplace data}

\item{taxonomy}{(\code{character}) path to the taxonomy file or a data frame
with two columns containing the taxonomy data; the first column contains
sequence IDs, while the second column contains comma-separated taxonomic
classifications.}

\item{outgroup}{(\code{character}) path to the outgroup file; or a character
vector containing the IDs of the outgroup sequences; or a character vector
containing the names of one or more taxa present in the taxonomy file.}

\item{ranks}{(\code{character}) vector of taxonomic ranks to use for the
assignment}

\item{ncpu}{(\code{integer}) number of threads to use}

\item{allow_file_overwriting}{(\code{logical}) whether to overwrite existing
output files}

\item{verbose}{(\code{logical}) whether to print verbose output}

\item{id_is_int}{(\code{logical}) whether the sequence IDs are integers}
}
\value{
(\code{data.frame}) with columns:
\itemize{
\item \code{seq_id} (character) the sequence ID
\item \code{rank} (ordered factor) the taxonomic rank
\item \code{parent_taxon} (character) the parent taxon
\item \code{taxon} (character) the taxon
\item \code{prob} (numeric) the probability of the assignment
}
}
\description{
Run Gappa
}
