% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krona.R
\name{generate_krona_data}
\alias{generate_krona_data}
\title{Generate data to write a KronaTools XML file}
\usage{
generate_krona_data(otu_taxonomy)
}
\arguments{
\item{otu_taxonomy}{(\code{data.frame}) with (at least) columns \code{kingdom}, \code{phylum},
\code{class}, \code{order}, \code{family}, \code{genus}, \code{species}, \code{nread} and \code{nsample}}
}
\value{
\code{data.frame} with columns \code{rank}, \code{taxon}, \code{parent_taxonomy},
\code{phylum_unknown_fread}, \code{phylum_unknown_fotu}, \code{phylum_unknown_focc},
\code{class_unknown_fread}, \code{class_unknown_fotu}, \code{class_unknown_focc},
\code{order_unknown_fread}, \code{order_unknown_fotu}, \code{order_unknown_focc},
\code{family_unknown_fread}, \code{family_unknown_fotu}, \code{family_unknown_focc},
\code{genus_unknown_fread}, \code{genus_unknown_fotu}, \code{genus_unknown_focc},
\code{species_unknown_fread}, \code{species_unknown_fotu}, \code{species_unknown_focc},
\code{nread}, \code{nocc}, \code{notu}, \code{child_unknown_fread}, \code{child_unknown_fotu},
\code{child_unknown_focc}, \code{fread}, \code{focc}, \code{fotu}
}
\description{
Generate data to write a KronaTools XML file
}
