% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{hmmalign}
\alias{hmmalign}
\title{Align query sequences to an HMM}
\usage{
hmmalign(
  seqs,
  hmm,
  outfile,
  outformat = "A2M",
  compress = endsWith(outfile, ".gz")
)
}
\arguments{
\item{seqs}{(\code{\link[Biostrings:XStringSet-class]{XStringSet}}, \code{character}
string giving a FASTA file name, \code{character} vector, or \code{data.frame})
sequences to align}

\item{hmm}{(\code{character} string giving a file name) HMM for alignment}

\item{outfile}{(\code{character} string) file name for output}

\item{outformat}{(\code{character} string) output format for alignment; options
are \code{"A2M"} and \code{"AFA"}, and lower-case versions of these.}

\item{compress}{(\code{logical}) if TRUE, compress the output file with gzip}
}
\value{
a \code{character} string giving the output file
}
\description{
Align query sequences to an HMM
}
