% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{large_preclosed_taxon_table}
\alias{large_preclosed_taxon_table}
\title{Calculate a pre-closed-reference table for taxa with many sequences}
\usage{
large_preclosed_taxon_table(
  known_taxon_table,
  asv_taxsort,
  rank,
  parent_rank,
  tax_ranks,
  min_ops = 1e+06
)
}
\arguments{
\item{known_taxon_table}{(\code{data.frame}) a table of known taxon assignments}

\item{asv_taxsort}{(\code{data.frame}) a table of ASV taxonomic assignments}

\item{rank}{(\code{character}) the taxonomic rank to calculate}

\item{parent_rank}{(\code{character}) the taxonomic rank to use as a parent}

\item{tax_ranks}{(\code{character}) the taxonomic ranks to use}

\item{min_ops}{(\code{integer} scalar) the minimum number of operations required
for a single parent taxon to be considered "large"}
}
\value{
(\code{data.frame}) a table of sequences for which closed-reference
clustering is required. This includes both sequences which are not assigned
to the target rank which will act as queries, and identified sequences
which belong to the same parent rank as one or more queries. The results
includes a \code{tar_group} column for dynamic branching over each parent
taxon. If there are no queries, a single row is returned to branch over.
}
\description{
This calculates a pre-closed-reference table for taxa with many sequences,
which may be efficiently parallelized within a single parent taxon.
}
