% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\docType{data}
\name{merged_filter_option_names}
\alias{merged_filter_option_names}
\title{Names of options in merged_filter_options}
\format{
An object of class \code{character} of length 5.
}
\usage{
merged_filter_option_names
}
\description{
Names of options in merged_filter_options
}
\keyword{datasets}
