% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{nhmmer}
\alias{nhmmer}
\title{Search for subsequences matching one or more nucleotide HMMs in a set of
sequences}
\usage{
nhmmer(seqs, hmm, ncpu = local_cpus())
}
\arguments{
\item{seqs}{(\code{\link[Biostrings:XStringSet-class]{XStringSet}}, \code{character}
file name of a FASTA file, \code{character} vector, or \code{data.frame}) sequences to search}

\item{hmm}{(\code{character} file name) path to HMM(s) to search for}

\item{ncpu}{(\code{integer}) number of threads to use for searching}
}
\value{
a \code{\link[tibble:tibble]{tibble}} listing the HMM hits
}
\description{
Search for subsequences matching one or more nucleotide HMMs in a set of
sequences
}
