% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{parse_project_name}
\alias{parse_project_name}
\alias{parse_custom_sample_table}
\alias{parse_parallel_options}
\alias{parse_forward_primer}
\alias{parse_reverse_primer}
\alias{parse_amplicon_model_options}
\alias{parse_control_options}
\alias{parse_taxonomy_options}
\alias{parse_taxonomy_ranks}
\alias{parse_protax_options}
\alias{parse_sintax_options}
\alias{parse_bayesant_options}
\alias{parse_epa_options}
\alias{parse_outgroup_options}
\title{Parse pipeline options from a YAML file}
\usage{
parse_project_name(pipeline_options)

parse_custom_sample_table(pipeline_options)

parse_parallel_options(pipeline_options)

parse_forward_primer(pipeline_options)

parse_reverse_primer(pipeline_options)

parse_amplicon_model_options(amplicon_model_options)

parse_control_options(pipeline_options)

parse_taxonomy_options(taxonomy_options)

parse_taxonomy_ranks(rank_options)

parse_protax_options(protax_options)

parse_sintax_options(sintax_options)

parse_bayesant_options(bayesant_options)

parse_epa_options(epa_options)

parse_outgroup_options(pipeline_options)
}
\arguments{
\item{pipeline_options}{(\code{list}) the pipeline options to parse}

\item{amplicon_model_options}{(\code{list}) the amplicon model options to parse}

\item{taxonomy_options}{(\code{list}) the taxonomy options to parse}

\item{rank_options}{(\code{list} or \code{character} vector) the taxonomic ranks to use in the
pipeline, in order from most inclusive (e.g., kingdom) to least inclusive
(e.g., species). Values may be either named or unnamed. When named, the name is
taken to be the rank, and the value is the "in-group" taxon at that rank, i.e.
the taxon for which results are desired. When unnamed, the value is taken to
be the rank. Example: \code{list(kingdom = "Fungi", "phylum", "class", "order", "family", "genus", "species")}}

\item{protax_options}{(\code{list}) the protax options to parse}

\item{sintax_options}{(\code{list}) the sintax options to parse}

\item{bayesant_options}{(\code{list}) the bayesant options to parse}

\item{epa_options}{(\code{list}) the epa options to parse}
}
\value{
\code{NULL}.  This function is called for its side effect, which is to
configure global options.
}
\description{
These functions parse the pipeline options from a YAML file and set
package options accordingly.
}
\keyword{internal}
