% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_options.R
\name{custom_sample_table}
\alias{custom_sample_table}
\alias{do_custom_sample_table}
\alias{max_workers}
\alias{min_workers}
\alias{workers_per_seqrun}
\alias{max_batchsize}
\alias{n_seqrun}
\alias{n_orient_seqrun}
\alias{n_workers}
\alias{forward_primer}
\alias{reverse_primer}
\alias{amplicon_model_type}
\alias{amplicon_model_file}
\alias{max_model_start}
\alias{min_model_end}
\alias{min_model_score}
\alias{do_model_filter}
\alias{do_model_align}
\alias{do_model_align_only}
\alias{do_model_filter_only}
\alias{do_model_both}
\alias{do_numt_filter}
\alias{spike_file}
\alias{do_spike}
\alias{pos_control_file}
\alias{do_pos_control}
\alias{do_protax}
\alias{protax_aligned}
\alias{protax_unaligned}
\alias{protax_location}
\alias{do_sintax}
\alias{sintax_ref}
\alias{do_bayesant}
\alias{bayesant_ref}
\alias{do_epa}
\alias{epa_ref}
\alias{epa_taxonomy}
\alias{epa_tree}
\alias{epa_params}
\alias{epa_outgroup}
\alias{do_outgroup}
\alias{parse_pipeline_options}
\title{Functions to access pipeline-wide options}
\usage{
custom_sample_table()

do_custom_sample_table()

max_workers()

min_workers()

workers_per_seqrun()

max_batchsize()

n_seqrun()

n_orient_seqrun()

n_workers()

forward_primer()

reverse_primer()

amplicon_model_type()

amplicon_model_file()

max_model_start()

min_model_end()

min_model_score()

do_model_filter()

do_model_align()

do_model_align_only()

do_model_filter_only()

do_model_both()

do_numt_filter()

spike_file()

do_spike()

pos_control_file()

do_pos_control()

do_protax()

protax_aligned()

protax_unaligned()

protax_location()

do_sintax()

sintax_ref()

do_bayesant()

bayesant_ref()

do_epa()

epa_ref()

epa_taxonomy()

epa_tree()

epa_params()

epa_outgroup()

do_outgroup()

parse_pipeline_options()
}
\value{
(\code{character} string) the forward primer

(\code{character} string) the reverse primer
}
\description{
When used in a targets plan, these should always be
pre-evaluated with \verb{!!} or \verb{!!!} to ensure proper dependency tracking.
}
