% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protax.R
\name{protax_besthit_closedref}
\alias{protax_besthit_closedref}
\title{Closed-reference single-linkage clustering of aligned sequences}
\usage{
protax_besthit_closedref(
  infile,
  index,
  i,
  unknowns,
  thresh,
  seq_width,
  ncpu = local_cpus(),
  max_gap = 100L
)
}
\arguments{
\item{infile}{(\code{character}) file name of gzipped FASTA file containing
sequences to cluster}

\item{index}{(\code{character}) file name of index for \code{infile}, as created by
\code{fastx_gz_index()}}

\item{i}{(\code{integer}) integer indices of sequences to cluster}

\item{unknowns}{(\code{logical}) logical vector indicating which sequences in \code{i}
are considered unknowns (the rest are references)}

\item{thresh}{(\code{numeric}) clustering similarity threshold as a percentage,
i.e., theoretically 0.0 to 100.0, but only values between 50.0 and 100.0 are
accepted, to avoid accidentally supplying distances (likely to be less than
50) or fractions (always between 0.0 and 1.0).}

\item{seq_width}{(\code{integer}) width of alignment}

\item{ncpu}{(\code{integer}) number of threads to use}

\item{max_gap}{(\code{integer}) passed to \code{fastx_gz_random_access_extract()}}
}
\value{
a \code{data.frame} with columns \code{seq_id}, \code{cluster}, and \code{dist}
}
\description{
Closed-reference single-linkage clustering of aligned sequences
}
