% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protax.R
\name{run_protax}
\alias{run_protax}
\title{Taxonomically identify sequences using the PERL implementation of ProtaxFungi}
\usage{
run_protax(
  seqs,
  outdir,
  modeldir,
  ncpu = local_cpus(),
  script = "scripts/runprotax"
)
}
\arguments{
\item{seqs}{(\code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, \code{character}
path to a FASTA file, \code{character} vector, or \code{data.frame}) sequences to
taxonomically identify}

\item{outdir}{(\code{character} string) directory to write output to. If the'
directory exists, all existing files in it will be deleted.}

\item{modeldir}{(\code{character} string) directory containing trained model files
for ProtaxFungi}

\item{ncpu}{(\code{integer}) number of threads to use for calls to USEARCH}

\item{script}{(\code{character} string) path to the ProtaxFungi script}
}
\value{
a \code{character} vector of output files generated by protax
}
\description{
Taxonomically identify sequences using the PERL implementation of ProtaxFungi
}
