% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protax.R
\name{run_protax_animal}
\alias{run_protax_animal}
\title{Taxonomically identify sequences using ProtaxAnimal}
\usage{
run_protax_animal(
  aln_seqs,
  modeldir,
  min_p = 0.1,
  rep_p = 0.01,
  strip_inserts = FALSE,
  id_is_int = FALSE,
  info = FALSE,
  options = character()
)
}
\arguments{
\item{aln_seqs}{(\code{character} vector) file name(s) of FASTA files containing
sequences to taxonomically identify. Sequences should all be aligned to the
reference alignment used to train Protax, for instance using \code{hmmalign()}.
If they end in \code{".gz"} then they are assumed to be compressed with gzip.}

\item{modeldir}{(\code{character} string) directory containing trained ProtaxA
model files.  These should include \code{"taxonomy.priors"}, \code{"refs.aln"},
\code{"model.rseqs.numeric"}, \code{"model.pars"}, \code{"model.scs"}.}

\item{min_p}{(\code{numeric}) minimum probability to recurse to the next rank}

\item{rep_p}{(\code{numeric}) minimum probability to include an identification in
the output}

\item{strip_inserts}{(\code{logical}) if \code{TRUE}, \code{aln_seqs} may include insertions
in lower-case, so strip these.}

\item{id_is_int}{(\code{logical}) if \code{TRUE}, sequence labels should be interpreted
as integers}

\item{info}{(\code{logical}) if \code{TRUE}, return information about which reference
sequences were the top two matches responsible for each identification}

\item{options}{(\code{character} vector) additional command line options to
Protax}
}
\value{
a \code{data.frame} with columns \code{seq_id}, \code{rank}, \code{taxonomy}, \code{prob},
and if \code{info} is \code{TRUE}, \code{best_id}, \code{best_dist}, \code{second_id}, \code{second_dist}
}
\description{
vectorized on aln_seqs
attempts to run them ALL in parallel, be careful!
}
