% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{sintax}
\alias{sintax}
\title{Perform taxonomic classification using SINTAX in VSEARCH}
\usage{
sintax(query, ref, ncpu = NULL, id_is_int = FALSE, hash = NULL)
}
\arguments{
\item{query}{(\code{data.frame}, \code{\link[Biostrings:XStringSet-class]{DNAStringSet}},
\code{character} vector, or file name) query sequences}

\item{ref}{(\code{data.frame}, \code{\link[Biostrings:XStringSet-class]{DNAStringSet}},
\code{character} vector, or file name) reference sequences}

\item{ncpu}{(\code{integer} count) number of threads to use}

\item{id_is_int}{(\code{logical} flag) if \code{TRUE}, return the sequence IDs as
integers}

\item{hash}{(\code{character} string) hash value for the queries; ignored (but
used by \code{targets} for dependency tracking)}
}
\value{
\code{tibble::tibble} with columns \code{seq_id} (or \code{seq_idx} if \code{id_is_int}
is TRUE), \code{rank}, \code{parent_taxonomy}, \code{taxon}, and \code{prob}, where \code{seq_id}
(\code{seq_idx}) is the ID of a sequence from \code{query}, \code{rank} is the taxonomic
rank of the taxon, \code{parent_taxonomy} is the parent taxonomic unit of the
taxon, \code{taxon} is the taxon name, and \code{prob} is the probability of the
taxon being the correct classification of the sequence
}
\description{
Perform taxonomic classification using SINTAX in VSEARCH
}
