% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{small_predenovo_taxon_table}
\alias{small_predenovo_taxon_table}
\title{Calculate a pre-denovo taxon table for taxa with few sequences}
\usage{
small_predenovo_taxon_table(
  closedref_taxon_table,
  asv_taxsort,
  rank,
  parent_rank,
  tax_ranks,
  max_ops = 1e+06
)
}
\arguments{
\item{closedref_taxon_table}{(\code{data.frame}) a table of sequences for which
closed-reference clustering is required. Must have columns \code{seq_idx}
(\code{integer}) giving the indices of sequences in \code{seq_file}, as well as
columns with names matching \code{rank} and \code{parent_rank}.}

\item{asv_taxsort}{(\code{data.frame}) a table of ASV taxonomic assignments}

\item{rank}{(\code{character}) the taxonomic rank to calculate}

\item{parent_rank}{(\code{character}) the taxonomic rank to use as a parent}

\item{tax_ranks}{(\code{character}) the taxonomic ranks to use}

\item{max_ops}{(\code{integer} scalar) the maximum number of operations to put
together in a single execution group}
}
\value{
(\code{data.frame}) a table of sequences for which denovo clustering is
required. This includes both sequences which are not assigned to the target
rank which will act as queries, and identified sequences which belong to the
same parent rank as one or more queries. The results includes a \code{tar_group}
column for dynamic branching over execution groups.
}
\description{
This calculates a pre-denovo taxon table for taxa with few sequences,
and groups them into approximately equally sized execution groups.
}
