% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{superranks}
\alias{superranks}
\alias{superrank_vars}
\alias{subranks}
\alias{subrank_vars}
\title{Get superordinate or subordinate taxonomic ranks}
\usage{
superranks(x = tip_rank(), ranks = tax_ranks())

superrank_vars(x = tip_rank(), ranks = tax_ranks())

subranks(x = ingroup_rank(), ranks = tax_ranks())

subrank_vars(x = ingroup_rank(), ranks = tax_ranks())
}
\arguments{
\item{x}{(\code{character} string) the taxonomic rank to find superordinate or
subordinate ranks for}

\item{ranks}{(\code{character} vector) the full list of taxonomic ranks, in order
from most inclusive to least inclusive}
}
\value{
a \code{character} vector of superordinate or subordinate taxonomic ranks

a \code{list} of \code{symbol}s representing the superordinate ranks

a \code{list} of \code{symbol}s representing the subordinate ranks
}
\description{
Get superordinate or subordinate taxonomic ranks
}
