% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{tax_ranks}
\alias{tax_ranks}
\alias{set_tax_ranks}
\title{Get or set the taxonomic ranks to use in the pipeline}
\usage{
tax_ranks()

set_tax_ranks(value)
}
\arguments{
\item{value}{(\code{character} vector) the taxonomic ranks to use in the pipeline,
in order from most inclusive (e.g., kingdom) to least inclusive (e.g., species)}
}
\value{
a \code{character} vector of taxonomic ranks

the previously set taxonomic ranks
}
\description{
Get or set the taxonomic ranks to use in the pipeline
}
