% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\name{trim_primer}
\alias{trim_primer}
\title{Trim primers from a set of sequences}
\usage{
trim_primer(seqs, primer, ...)
}
\arguments{
\item{seqs}{(\code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, \code{character}
vector, or \code{data.frame}) sequences to trim}

\item{primer}{(\code{character}) primer sequence}

\item{...}{additional arguments to \code{cutadapt_filter_trim()}}
}
\value{
a \code{tibble} with columns \code{seq_id} and \code{seq}
}
\description{
Trim primers from a set of sequences
}
