% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\name{trim_raw_pairs}
\alias{trim_raw_pairs}
\title{Top-level function to trim a batch of raw paired-end reads}
\usage{
trim_raw_pairs(
  pairs_meta,
  seqrun,
  orient,
  trim_options,
  primer_R1,
  primer_R2,
  raw_path = "sequences/01_raw",
  ncpu = local_cpus()
)
}
\arguments{
\item{pairs_meta}{(\code{data.frame}) metadata for the paired-end reads}

\item{seqrun}{(\code{character}) seqrun identifier for the batch}

\item{orient}{(\code{character}) orientation of the reads for the batch; either
"fwd" or "rev"}

\item{trim_options}{(\href{`cutadapt_paired_options()`}{\code{cutadapt_paired_options}}
options for cutadapt}

\item{primer_R1}{(\code{character}) primer specification for R1}

\item{primer_R2}{(\code{character}) primer specification for R2}

\item{raw_path}{(\code{character}) path to the raw reads}

\item{ncpu}{(\code{integer}) number of CPU threads to use}
}
\value{
a \code{character} vector with the path to the trimmed output files
}
\description{
Top-level function to trim a batch of raw paired-end reads
}
