% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{vsearch_cluster_smallmem}
\alias{vsearch_cluster_smallmem}
\title{"cluster_smallmem" function of vsearch}
\usage{
vsearch_cluster_smallmem(seq, threshold = 1, ncpu = local_cpus())
}
\arguments{
\item{seq}{(\code{data.frame}, \code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, \code{character} vector, or
file name) sequences to cluster}

\item{threshold}{(\code{numeric} scalar) identity threshold, in range 0.0-1.0}

\item{ncpu}{(\code{integer} count) number of threads to use}
}
\value{
\code{tibble::tibble} with columns \code{query} and \code{hit}, where \code{query} is
the name of a sequence from \code{seq}, and \code{hit} is the name of the sequence
which is the centroid of the cluster containing \code{query}
}
\description{
"cluster_smallmem" function of vsearch
}
