% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{vsearch_uchime_ref}
\alias{vsearch_uchime_ref}
\title{"uchime_ref" function of vsearch}
\usage{
vsearch_uchime_ref(
  query,
  ref,
  ncpu = local_cpus(),
  id_only = FALSE,
  id_is_int = FALSE
)
}
\arguments{
\item{query}{(\code{data.frame}, \code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, \code{character} vector,
or file name) query sequences}

\item{ref}{(\code{data.frame}, \code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, \code{character} vector, or
file name) reference sequences}

\item{ncpu}{(\code{integer} count) number of threads to use}

\item{id_only}{(\code{logical} flag) if \code{TRUE}, return only the sequence IDs}

\item{id_is_int}{(\code{logical} flag) if \code{TRUE}, return the sequence IDs as
integers}
}
\value{
if \code{id_only} is FALSE, a \code{tibble::tibble} with columns \code{seq_id} (or
\code{seq_idx} if \code{id_is_int} is TRUE) and \code{seq}, where \code{seq_id} (\code{seq_idx}) is
the name of a sequence from \code{query} that is a chimera, and \code{seq} is the
sequence itself.  Alternatively, if \code{id_only} is TRUE, a \code{character} vector
(or \code{integer} vector if \code{id_is_int} is TRUE) of the sequence IDs which are
chimeras.
}
\description{
"uchime_ref" function of vsearch
}
