% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch.R
\name{vsearch_usearch_global}
\alias{vsearch_usearch_global}
\title{"usearch_global" function of vsearch}
\usage{
vsearch_usearch_global(
  query,
  ref,
  threshold,
  global = TRUE,
  ncpu = local_cpus(),
  id_is_int = FALSE
)
}
\arguments{
\item{query}{(\code{data.frame}, \code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, \code{character} vector, or
file name) query sequences}

\item{ref}{(\code{data.frame}, \code{\link[Biostrings:XStringSet-class]{DNAStringSet}}, \code{character} vector, or
file name) reference sequences}

\item{threshold}{(\code{numeric} scalar) identity threshold, in range 0.0-1.0}

\item{global}{(\code{logical} flag) if \code{TRUE}, end gaps and internal gaps are
penalized equally.  Otherwise end gaps are not penalized.}

\item{ncpu}{(\code{integer} count) number of threads to use}

\item{id_is_int}{(\code{logical} flag) if \code{TRUE}, return the sequence IDs as
integers}
}
\value{
\code{tibble::tibble} with columns \code{seq_id}, \code{clust}, and \code{dist}, where \code{seq_id} is
the name of a sequence from \code{query}, \code{clust} is the closest match to that
sequence in \code{ref}, and \code{dist} is the distance between them
}
\description{
"usearch_global" function of vsearch
}
