% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_and_return_file.R
\name{write_and_return_file}
\alias{write_and_return_file}
\alias{write_and_return_file.XStringSet}
\alias{write_and_return_file.data.frame}
\alias{write_and_return_file.character}
\alias{write_and_return_file.ggplot}
\alias{write_and_return_file.default}
\title{Write an object to a file and return the file path}
\usage{
write_and_return_file(x, file, ...)

\method{write_and_return_file}{XStringSet}(x, file, width = 20001L, ...)

\method{write_and_return_file}{data.frame}(x, file, type = c("rds", "tsv"), ...)

\method{write_and_return_file}{character}(x, file, ...)

\method{write_and_return_file}{ggplot}(x, file, ...)

\method{write_and_return_file}{default}(x, file, ...)
}
\arguments{
\item{x}{(any object) object to write}

\item{file}{(\code{character} string) file path to write to}

\item{...}{additional arguments to pass to the writing function}

\item{width}{(integer) the maximum width of each line in the output file}

\item{type}{(\code{character} string) the type of file to write to; currently
supported are \code{"rds"} (default) and \code{"tsv"}}
}
\description{
Write an object to a file and return the file path
}
